% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-decomposition.R
\name{ams.decomposition}
\alias{ams.decomposition}
\title{Decompose Annual Maximum Streamflow Data}
\usage{
ams.decomposition(data, years, trend)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as non-stationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\value{
List; decomposed AMS data with the same names as argument \code{ams}.
}
\description{
Removes trends in the mean and/or variance of annual maximum streamflow (AMS)
data using Sen’s trend estimator and/or a moving‐window estimator of the variance.
Four scenarios are supported:
\enumerate{
\item No trend. The data is returned unmodified.
\item Trend in the mean only.
\item Trend in the variance only.
\item Trends in both the mean and the variance.
}
}
\details{
Internally, the function does the following:
\enumerate{
\item Compute \code{covariate = (ams$year - 1900) / 100}.
\item If there is a trend in the location, fit Sen’s trend estimator to
\code{ams$data} and \code{covariate}. Then, remove the fitted linear trend.
\item If there is a trend in the scale, compute the moving‐window variance using
\link{mw_variance}, fit Sen’s trend estimator to those variances, and then
rescales the series to remove trends in the scale.
\item If necessary, shift the data so that its minimum is at least 1.
}
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
ams.decomposition(data, years, trend)

}
\seealso{
\link{mw_variance}, \link{sens_trend}
}
