% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-mw-variance.R
\name{ams_mw_variance}
\alias{ams_mw_variance}
\title{Estimate Variance for Annual Maximum Streamflow Data}
\usage{
ams_mw_variance(data, years, size = 10L, step = 5L)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{size}{Integer scalar. The number of years in each moving window.
Must be a positive number less than or equal to \code{length(data)}
(default is 10). If \code{length(data) < size}, an error is raised.}

\item{step}{Integer scalar. The offset (in years) between successive
moving windows. Must be a positive number (default is 5).}
}
\value{
A list with two entries:
\itemize{
\item \code{years}: Numeric vector containing the mean year within each window.
\item \code{std}: Numeric vector of standard deviations within each window.
}
}
\description{
This function estimates the standard deviation of a vector of annual maximum
streamflow (AMS) data using a moving window algorithm, returning a list that
pairs each window’s mean year with its computed standard deviation. The
parameters \code{size} and \code{step} parameters control the behaviour of the window.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
ams_mw_variance(data, years)

}
