% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-05BA001.R
\docType{data}
\name{CAN_05BA001}
\alias{CAN_05BA001}
\title{CAN-05BA001}
\format{
A dateframe with 62 rows and 2 columns spanning the period 1913-2023.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_05BA001
}
\description{
A dataframe of annual maximum series observations for
station 05BA001, BOW RIVER AT LAKE LOUISE in Alberta, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the annual maximum series observation, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding year.
}
}
\section{Additional Information}{

This is an \emph{unregulated} station that is not in the RHBN network. Other notable features include:
\itemize{
\item There are no observations from 1919-1965. Missing data should be handled carefully.
\item The MKS/Pettitt tests do not find evidence of change points at the 0.05 significance level.
\item Trend detection finds evidence of a trend in variability.
\item If nonstationarity is assumed, RFPL uncertainty quantification fails on this dataset.
}

This dataset is used as a test case for failure modes in RFPL and variability estimation.
}

\keyword{datasets}
