% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-07BE001.R
\docType{data}
\name{CAN_07BE001}
\alias{CAN_07BE001}
\title{CAN-07BE001}
\format{
A dateframe with 108 rows and 2 columns spanning the period 1913-2020.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_07BE001
}
\description{
A dataframe of annual maximum series observations for
station 07BE001, ATHABASCA RIVER AT ATHABASCA in Alberta, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the annual maximum series observation, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding year.
}
}
\section{Additional Information}{

This is an unregulated station that is not in the RHBN. Additionally,
\itemize{
\item The MKS/Pettitt tests find no evidence of change points at the 0.05 significance level.
\item Trend detection finds no evidence of trends in the mean or variability.
}

This dataset is an excellent introductory example to stationary FFA.
}

\keyword{datasets}
