% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-08MH016.R
\docType{data}
\name{CAN_08MH016}
\alias{CAN_08MH016}
\title{CAN-08MH016}
\format{
A dateframe with 95 rows and 2 columns spanning the period 1922-2016.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_08MH016
}
\description{
A dataframe of annual maximum series observations for
station 08MH016, CHILLIWACK RIVER AT CHILLIWACK LAKE in British Columbia, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the annual maximum series observation, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding year.
}
}
\section{Additional Information}{

This is an unregulated station in the RHBN. Additionally,
\itemize{
\item The MKS/Pettitt tests find no evidence of change points at the 0.05 significance level.
\item Trend detection finds evidence of an increasing trend in the variability.
}
}

\keyword{datasets}
