% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-08NH021.R
\docType{data}
\name{CAN_08NH021}
\alias{CAN_08NH021}
\title{CAN-08NH021}
\format{
A dataframe with 91 rows and 2 columns spanning the period 1928-2018.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_08NH021
}
\description{
A dataframe of annual maximum series observations for
station 08NH021, KOOTENAI RIVER AT PORTHILL in British Columbia, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the observed annual maximum series, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding years.
}
}
\section{Additional Information}{

This is a regulated station that is not in the RHBN. Additionally,
\itemize{
\item The Libby dam was constructed upstream of this station in 1972.
\item The Pettitt test finds evidence of a change point in 1972 at the 0.05 significance level.
\item The MKS test finds evidence of change points in 1960 & 1985 at the 0.05 significance level.
}

This dataset is an excellent introduction to change point detection and piecewise NS-FFA.
}

\keyword{datasets}
