% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-08NM116.R
\docType{data}
\name{CAN_08NM116}
\alias{CAN_08NM116}
\title{CAN-08NM116}
\format{
A dateframe with 75 rows and 2 columns spanning the period 1949-2023.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_08NM116
}
\description{
A dataframe of annual maximum series observations for
station 08NM116, MISSION CREEK NEAR EAST KELOWNA in British Columbia, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the annual maximum series (AMS) observation, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding year.
}
}
\section{Features}{

\itemize{
\item The remoteness of this station makes it suitable for flood frequency analysis.
\item The MKS/Pettitt tests do not find evidence of change points at the 0.05 significance level.
\item Trend detection finds evidence of a trend in variability.
}

This dataset is an excellent hands-on application of the FFA framework for practitioners.
}

\keyword{datasets}
