% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-kpss-test.R
\name{eda_kpss_test}
\alias{eda_kpss_test}
\title{Kwiatkowski–Phillips–Schmidt–Shin (KPSS) Unit Root Test}
\usage{
eda_kpss_test(data, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
A list containing the test results, including:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{statistic}: The KPSS test statistic.
\item \code{p_value}: The interpolated p-value. See note regarding discrete thresholds.
\item \code{reject}: Logical scalar. If, TRUE the null hypothesis is rejected at \code{alpha}.
\item \code{msg}: Character string summarizing the test outcome, printed if \code{quiet = FALSE}.
}
}
\description{
Performs the KPSS unit root test on annual maximum series (AMS) data.
The null hypothesis is that the time series is trend-stationary with a linear
trend and constant drift. The alternative hypothesis is that the time series
has a unit root and is nonstationary.
}
\details{
The implementation of the KPSS test is based on the \pkg{aTSA} package, which
interpolates a significance table from Hobjin et al. (2004). Therefore, a result
of \eqn{p = 0.01} implies that \eqn{p \leq 0.01} and a result of \eqn{p = 0.10}
implies that \eqn{p \geq 0.10}. This implementation uses the Type III KPSS test,
which accounts for a linear trend in the data.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
eda_kpss_test(data)

}
\references{
Hobijn, B., Franses, P.H. and Ooms, M. (2004), Generalizations of the KPSS-test
for stationarity. Statistica Neerlandica, 58: 483-502.
}
\seealso{
\link{eda_pp_test}
}
