% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-pp-test.R
\name{eda_pp_test}
\alias{eda_pp_test}
\title{Phillips–Perron Unit Root Test}
\usage{
eda_pp_test(data, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
List; the test results, consisting of:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{statistic}: The Z-statistic used to perform the test.
\item \code{p.value}: Reported p-value from the test. See notes on interpolation thresholds.
\item \code{reject}: Logical. If \code{TRUE}, the null hypothesis was rejected at \code{alpha}.
\item \code{msg}: Character string summarizing the test result, printed if \code{quiet = FALSE}.
}
}
\description{
Applies the Phillips–Perron (PP) test to check for a unit root in annual
maximum streamflow (AMS) data. The null hypothesis is that the series contains a
unit root (and is thus non-stationary). This implementation of the PP test assumes
the time series has both stationary drift and a linear trend.
}
\details{
The implementation of this test is based on the \pkg{aTSA} package, which
interpolates p-values from a table of critical values presented in Fuller W. A.
(1996). The critical values are only available for \eqn{\alpha \geq 0.01}.
Therefore, a reported p-value of 0.01 indicates \eqn{p \leq 0.01}.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
eda_pp_test(data)

}
\references{
Fuller, W. A. (1996). Introduction to statistical time series, second ed., Wiley,
New York.
}
\seealso{
\link{eda_kpss_test}
}
