% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-package.R
\docType{package}
\name{ffaframework-package}
\alias{ffaframework}
\alias{ffaframework-package}
\title{Flood Frequency Analysis Framework}
\description{
This package provides tools for stationary (S-FFA) and nonstationary (NS-FFA)
flood frequency analysis of annual maximum series data. The package supports
flexible workflows with the following core components:
}
\details{
\enumerate{
\item \strong{Test} for evidence of nonstationarity in annual maximum series data to
inform approach selection (S-FFA or NS-FFA) using the \verb{eda_*} methods:
\itemize{
\item Detect statistically significant change points.
\item Detect statistically significant temporal trends in the \emph{mean} and \emph{variance}.
}
\item \strong{Select} a suitable probability distribution using the \verb{select_* } methods.
\item \strong{Fit} model parameters using the \verb{fit_*} methods.
\item \strong{Quantify} uncertainty using the \verb{uncertainty_*} methods.
\item \strong{Evaluate} model performance using \code{model_diagnostics()}.
}

Additional utility functions support visualization and computation:
\itemize{
\item \verb{plot_*} methods produce diagnostic and summary plots.
\item \verb{data_*} methods load, transform, and decompose annual maximum series data.
\item \verb{quantile_*}, \verb{loglik_*}, and \verb{lmom_*} implement distribution-specific computations.
}

This package assumes familiarity with statistical techniques used in FFA, including
L-moments, maximum likelihood estimation, and parametric bootstrap.
For an explanation of these methods, see the
\href{https://rileywheadon.github.io/ffa-docs/}{FFA Framework wiki}.
For examples, see the vignettes on exploratory data analysis and flood frequency
analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rileywheadon.github.io/ffa-docs/}
  \item Report bugs at \url{https://github.com/rileywheadon/ffa-framework/issues}
}

}
\author{
\strong{Maintainer}: Riley Wheadon \email{rileywheadon@gmail.com}

Authors:
\itemize{
  \item Cuauhtémoc Vidrio-Sahagún \email{ct.vidrio-sahagun@usask.ca}
  \item Alain Pietroniro \email{alain.pietroniro@ucalgary.ca}
  \item Jianxun He \email{jianhe@ucalgary.ca}
}

}
\keyword{package}
