% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework.R
\name{ffaframework}
\alias{ffaframework}
\title{Flood Frequency Analysis Framework}
\usage{
ffaframework(
  data,
  years,
  splits = NULL,
  structures = NULL,
  automatic = FALSE,
  ...
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{splits}{An integer vector of years at which to split the data (default is \code{NULL}).
A split point is the \emph{first} year in a subperiod.}

\item{structures}{An list of structure objects of size \code{length(splits) + 1} (default is \code{NULL}).
Each structure object is a list with boolean items \code{location} and \code{scale} indicating a trend
in the mean/variability respectively.}

\item{automatic}{If \code{TRUE}, the split points and nonstationary structures are chosen
automatically using the results of EDA (default is \code{FALSE}). This argument is ignored
if both \code{splits} and \code{structures} are not \code{NULL}.}

\item{...}{Additional arguments to be passed to the statistical tests and frequency
analysis functions. See the details of \code{\link[=module_eda]{module_eda()}} and \code{\link[=module_ffa]{module_ffa()}} for a complete
list.}
}
\value{
\code{summary}: A list containing high-level information about the results:
\itemize{
\item \code{splits}: The vector of years used to split the data into subperiods.
\item \code{structures}: A list of nonstationary structure objects with size \code{length(splits) + 1}.
}

\code{blocks}: A list of results for each module. Each block is a list containing:
\itemize{
\item \code{name}: Either \code{"Change Points"}, \code{"Trend Detection"}, or \code{"Frequency Analysis"}.
\item \code{start}: The first year of the subperiod.
\item \code{end}: The last year of the subperiod.
\item Additional items depending on the module.
}
}
\description{
Runs the entire flood frequency analysis framework using the exploratory
data analysis (\code{\link[=module_eda]{module_eda()}}) and flood frequency analysis (\code{\link[=module_ffa]{module_ffa()}}) modules.
Returns a comprehensive and reproducible summary of the results.
}
\details{
The behaviour of the function when \code{automatic = FALSE} can be confusing. If either
of \code{splits} or \code{structures} is \code{NULL}, the user will be prompted to select them. However,
if R is running in non-interactive mode, this is not possible, so the function will return
all the results up to the current block along with an error message.

To avoid this issue, the user has three options:
\enumerate{
\item Set \code{splits} and \code{structures} manually.
\item Run R in interactive mode (using RStudio, for example).
\item Set \code{automatic = TRUE}.
}
}
