% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-fast.R
\name{fit_lmom_fast}
\alias{fit_lmom_fast}
\title{Helper Function for L-moments Parameter Estimation}
\usage{
fit_lmom_fast(data, distribution)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{distribution}{Character scalar. A three-character code indicating
a distribution family. Must be one of: \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}
}
\value{
A list containing the results of parameter estimation:
\itemize{
\item \code{method}: \code{"L-moments"}.
\item \code{params}: numeric vector of 2 or 3 parameters depending on the distribution.
}
}
\description{
A helper function used by \code{\link[=fit_lmom_xxx]{fit_lmom_xxx()}}.
This function does not validate parameters and is designed for use in other methods.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmom_fast(data, "PE3")

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\code{\link[=lmom_sample]{lmom_sample()}}, \code{\link[=fit_lmom_kappa]{fit_lmom_kappa()}}, \code{\link[=fit_lmom_xxx]{fit_lmom_xxx()}}
}
