% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-kappa.R
\name{fit_lmom_kappa}
\alias{fit_lmom_kappa}
\title{L-Moments Parameter Estimation for the Kappa Distribution}
\usage{
fit_lmom_kappa(data)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}
}
\value{
A list containing the results of parameter estimation:
\itemize{
\item \code{method}: \code{"L-moments"}.
\item \code{params}: numeric vector of 4 parameters in the order location, scale, shape (2).
}
}
\description{
This functions estimates the parameters of the four-parameter Kappa distribution
using the method of L-moments. Since there is no known closed form solution for
the parameters in terms of the L-moments, the parameters are computed numerically
using Newton-Raphson iteration.
}
\details{
First, the sample L-moments of the data are computed using the \code{\link[=lmom_sample]{lmom_sample()}}
method. Then, the \code{\link[stats:optim]{stats::optim()}} function is used to determine the parameters
by minimizing the euclidian distance between the sample and theoretical L-moment
ratios. The implementation of this routine is based on the deprecated \code{homtest}
package, formerly avilable at \url{https://CRAN.R-project.org/package=homtest}.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmom_kappa(data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\code{\link[=lmom_sample]{lmom_sample()}}, \code{\link[=fit_lmom_fast]{fit_lmom_fast()}}, \code{\link[=fit_lmom_xxx]{fit_lmom_xxx()}}
}
