% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-maximum-likelihood.R
\name{fit_maximum_likelihood}
\alias{fit_maximum_likelihood}
\title{Maximum Likelihood Parameter Estimation}
\usage{
fit_maximum_likelihood(
  data,
  distribution,
  prior = NULL,
  years = NULL,
  structure = NULL
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{distribution}{Character scalar. A three-character code indicating
a distribution family. Must be one of: \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{distribution = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}
}
\value{
A list containing the results of parameter estimation:
\itemize{
\item \code{params}: Numeric vector of estimated parameters.
\item \code{mll}: Maximum log‐likelihood value.
}
}
\description{
Estimates parameters of a probability distribution with an optional nonstationary
structure by  maximizing the log‐likelihood. Initial values are obtained through
L‐moment parameter estimation, and optimization is performed via \code{\link[stats:nlminb]{stats::nlminb()}}
with repeated perturbations if needed.
}
\details{
\enumerate{
\item Calls \code{\link[=fit_lmom_fast]{fit_lmom_fast()}} on \code{data} to obtain initial parameter estimates.
\item Initializes trend parameters to zero if necessary.
\item For \code{WEI} models, sets the location parameter to zero to ensure support.
\item Defines an objective function using \code{\link[=loglik_fast]{loglik_fast()}} or \code{\link[=general_loglik_fast]{general_loglik_fast()}}.
\item Runs \code{\link[stats:nlminb]{stats::nlminb()}} with box constraints. Attempts optimization
up to 100 times if a maximum cannot be found.
}
}
\note{
Although the more modern \code{\link[stats:optim]{stats::optim()}} function is preferred over
\code{\link[stats:nlminb]{stats::nlminb()}}, we use \code{\link[stats:nlminb]{stats::nlminb()}} because it supports infinite
values of the likelihood function.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
structure <- list(location = TRUE, scale = FALSE)
fit_maximum_likelihood(data, "GNO", NULL, years, structure)

}
\seealso{
\code{\link[=loglik_fast]{loglik_fast()}}, \code{\link[=general_loglik_fast]{general_loglik_fast()}}, \code{\link[=fit_lmom_fast]{fit_lmom_fast()}},
\code{\link[stats:nlminb]{stats::nlminb()}}
}
