% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-eda.R
\name{framework_eda}
\alias{framework_eda}
\title{Orchestrate Exploratory Data Analysis}
\usage{
framework_eda(data, years, ns_splits = NULL, automatic = FALSE)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{ns_splits}{An integer vector of years at which to split the data (default is \code{NULL}).
A split point is the \emph{first} year in a subperiod.}

\item{automatic}{If \code{TRUE}, the data is split automatically using the results of the Pettitt
and MKS tests (default is \code{FALSE}). This argument is ignored if \code{ns_splits} is not \code{NULL}.}

\item{...}{Additional arguments. See the "Optional Arguments" section for a complete list.}
}
\value{
\code{recommendations}: A list of recommended split points and nonstationary structures:
\itemize{
\item \code{approach}: Either \code{"S-FFA"}, \code{"NS-FFA"}, or \code{"Piecewise NS-FFA"}.
\item \code{splits}: The change point(s) identified by the change point detection test with the
the lowest statistically significant p-value, or an empty vector if no such test exists.
\item \code{structures}: A list of nonstationary structure objects for each subperiod. Each structure
is a list with boolean items \code{location} and \code{scale} which represent a linear trend in the
in the mean or variability of the data respectively.
}

\code{submodules}: A list of lists of statistical tests. Each sublist contains:
\itemize{
\item \code{name}: Either \code{"Change Points"} or \code{"Trend Detection"}.
\item \code{start}: The first year of the subperiod.
\item \code{end}: The last year of the subperiod.
\item Additional items from the statistical tests within the submodule.
}
}
\description{
First, this method identifies change points in annual maximum series data. Then, the
user is given the option to split the dataset into two or more subperiods. Finally,
this method performs a collection of statistical tests for identifying nonstationarity
in the mean and variability of each subperiod.
}
\details{
If the \code{ns_splits} argument is not \code{NULL}, it will be used to split the data.
To avoid splitting the data, use \code{splits = integer(0)}. Note that \code{integer(0)} is not
the same as \code{NULL}.

If \code{ns_splits} is \code{NULL} and \code{automatic} is \code{TRUE}, the data will be split automatically
based on the change point detection test with the lowest statistically significant p-value.
If neither test yielded a statistically significant p-value, the data will not be split.

If \code{ns_splits} is \code{NULL} and \code{automatic} is \code{FALSE}, the user will be prompted to select the
split points manually. This feature requires R to be running in interactive mode. Check if
R is in interactive mode using the built-in \code{interactive()} function.
}
\section{Optional Arguments}{

\itemize{
\item \code{alpha}: The numeric significance level for all statistical tests (default is 0.05).
\item \code{bbmk_samples}: The number of samples used in the Block-Bootstrap Mann-Kendall (BBMK)
test (default is 10000). Must be an integer.
}
}

\seealso{
\code{\link[=eda_pettitt_test]{eda_pettitt_test()}}, \code{\link[=eda_mks_test]{eda_mks_test()}}, \code{\link[=eda_mk_test]{eda_mk_test()}}, \code{\link[=eda_spearman_test]{eda_spearman_test()}}
\code{\link[=eda_bbmk_test]{eda_bbmk_test()}}, \code{\link[=eda_pp_test]{eda_pp_test()}}, \code{\link[=eda_kpss_test]{eda_kpss_test()}}, \code{\link[=eda_sens_trend]{eda_sens_trend()}},
\code{\link[=eda_runs_test]{eda_runs_test()}}, \code{\link[=eda_white_test]{eda_white_test()}}
}
