% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglik-fast.R
\name{general_loglik_fast}
\alias{general_loglik_fast}
\title{Generalized Log-Likelihood Helper Function}
\usage{
general_loglik_fast(data, distribution, params, prior, years, structure)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{distribution}{A three-character code indicating the distribution family.
Must be \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"},
or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}

\item{prior}{Numeric vector of length 2. Specifies the parameters of the
Beta prior for the shape parameter \eqn{\kappa}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}
}
\value{
Numeric scalar. The generalized log-likelihood value.
}
\description{
A helper function used by \code{\link[=general_loglik_gev]{general_loglik_gev()}}.
This function does not validate parameters and is designed for use in other methods.
}
\examples{
# Initialize data, params, prior, years, and structure
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0, 0)
prior <- c(5, 10)
years <- seq(from = 1901, to = 2000)
structure <- list(location = TRUE, scale = TRUE)

# Compute the generalized log-likelihood
general_loglik_fast(data, "GEV", params, prior, years, structure)

}
\seealso{
\code{\link[=general_loglik_gev]{general_loglik_gev()}}.
}
