% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglik-methods.R
\name{general_loglik_gev}
\alias{general_loglik_gev}
\title{Generalized Log-Likelihood Functions for GEV Models}
\usage{
general_loglik_gev(data, params, prior, years = NULL, trend = NULL)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{model = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\value{
Numeric scalar. The generalized log-likelihood value.
}
\description{
Computes the generalized log-likelihood for stationary and nonstationary
variants of the  Generalized Extreme Value (GEV) distribution with a Beta
prior on the shape parameter.
}
\details{
The generalized log-likelihood is defined as sum of the log-likelihood of the
specified model and the log-density of the Beta prior with parameters \eqn{(p, q)}.
The contribution of the prior is: \deqn{\log \pi(\kappa) = (p-1) \log(0.5-\kappa) 
+ (q-1) \log(0.5+\kappa) - \log (\beta(p, q))}
}
\examples{
# Initialize data, params, prior, years, and trend
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0)
prior <- c(5, 10)
years <- seq(from = 1901, to = 2000)
trend <- list(location = TRUE, scale = FALSE)

# Compute the generalized log-likelihood
general_loglik_gev(data, params, prior, years, trend)

}
\seealso{
\link{general_loglik_fast}
}
