% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gll-xxx.R
\name{gllxxx}
\alias{gllxxx}
\title{Generalized Log-Likelihood Helper Function}
\usage{
gllxxx(name, signature, data, params, prior, covariate = NULL)
}
\arguments{
\item{name}{Character (1); the name of the probability distribution.}

\item{signature}{Character (1); the non-stationary signature (\code{NULL}, \code{"10"}, or \code{"11"}).}

\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{prior}{Numeric (2); a vector of parameters \eqn{(p, q)} of the Beta prior
on \eqn{\kappa}.}

\item{covariate}{Numeric; a vector with the same length as \code{data}.
Required if \code{signature} is \code{"10"} or \code{"11"}.}
}
\value{
Numeric (1); the generalized log-likelihood value.
}
\description{
A helper function used by \link{gllgev}, \link{gllgev100}, and \link{gllgev110}.
}
\note{
This function does not perform parameter validation, which improves performance
but may cause unpredictable behaviour. Use at your own risk.
}
\examples{
# Initialize data, params, params, prior, and covariate
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0, 0)
prior <- c(5, 10)
covariate <- seq(0, 1, length.out = 100)

# Compute the generalized log-likelihood
gllxxx("GEV", "11", data, params, prior, covariate)

}
\seealso{
\link{gll-functions}
}
