% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik-methods.R
\name{llv-functions}
\alias{llv-functions}
\alias{llvgum}
\alias{llvgum10}
\alias{llvgum11}
\alias{llvnor}
\alias{llvnor10}
\alias{llvnor11}
\alias{llvlno}
\alias{llvlno10}
\alias{llvlno11}
\alias{llvgev}
\alias{llvgev100}
\alias{llvgev110}
\alias{llvglo}
\alias{llvglo100}
\alias{llvglo110}
\alias{llvgno}
\alias{llvgno100}
\alias{llvgno110}
\alias{llvpe3}
\alias{llvpe3100}
\alias{llvpe3110}
\alias{llvlp3}
\alias{llvlp3100}
\alias{llvlp3110}
\alias{llvwei}
\alias{llvwei100}
\alias{llvwei110}
\title{Log-Likelihood Functions for Probability Models}
\usage{
llvgum(data, params, years = NULL)

llvgum10(data, params, years)

llvgum11(data, params, years)

llvnor(data, params, years = NULL)

llvnor10(data, params, years)

llvnor11(data, params, years)

llvlno(data, params, years = NULL)

llvlno10(data, params, years)

llvlno11(data, params, years)

llvgev(data, params, years = NULL)

llvgev100(data, params, years)

llvgev110(data, params, years)

llvglo(data, params, years = NULL)

llvglo100(data, params, years)

llvglo110(data, params, years)

llvgno(data, params, years = NULL)

llvgno100(data, params, years)

llvgno110(data, params, years)

llvpe3(data, params, years = NULL)

llvpe3100(data, params, years)

llvpe3110(data, params, years)

llvlp3(data, params, years = NULL)

llvlp3100(data, params, years)

llvlp3110(data, params, years)

llvwei(data, params, years = NULL)

llvwei100(data, params, years)

llvwei110(data, params, years)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.
Required for non-stationary models, which end in \code{10}, \code{11}, \code{100}, or \code{110}.}
}
\value{
Numeric (1); the log-likelihood value.
}
\description{
Compute the log-likelihood value for stationary and non-stationary variants
of nine different distributions (\code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO}, \code{GNO},
\code{PE3}, \code{LP3}, and \code{WEI}).
}
\details{
The log-likelihood is the logarithm of the probability density function.
For two-parameter distributions (\code{GUM}, \code{NOR}, \code{LNO}), there are three
different \code{llv} functions:
\itemize{
\item \code{llv...()}: Stationary location and scale, 2 parameters.
\item \code{llv...10()}: Time-varying location, stationary scale, 3 parameters.
\item \code{llv...11()}: Time-varying location and scale, 4 parameters.
}

For three-parameter distributions (\code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, \code{WEI}),
there are also three different \code{llv} functions:
\itemize{
\item \code{llv...()}: Stationary location and scale, 3 parameters.
\item \code{llv...100()}: Time-varying location, stationary scale, 4 parameters.
\item \code{llv...110()}: Time-varying location and scale, 5 parameters.
}
}
\note{
The \code{llv...} functions perform extensive parameter validation, which can be slow.
If you plan to make calls these methods often, it is recommended to use the \link{llvxxx}
helper function instead.
}
\examples{
# Initialize data, years, and params
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
params <- c(0, 1, 1, 0)

# Compute the log-likelihood
llvgno100(data, params, years)

}
\seealso{
\link{llvxxx}
}
