% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik-fast.R
\name{llvxxx}
\alias{llvxxx}
\title{Log-Likelihood Helper Function}
\usage{
llvxxx(model, trend, data, params, years = NULL)
}
\arguments{
\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{covariate}{Numeric; a vector with the same length as \code{data}.
Required if \code{trend$location} or \code{trend$scale} is TRUE.}
}
\value{
Numeric (1); the log-likelihood value.
}
\description{
A helper function used by \link{llv-functions}.
}
\note{
This function does not perform parameter validation, which improves performance
at the cost of potentially unpredictable behaviour. Use at your own risk.
}
\examples{
# Initialize data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 0)

# Compute the log-likelihood
llvxxx("GEV", NULL, data, params)

}
\seealso{
\link{llv-functions}
}
