% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom-fast.R
\name{lmom_fast}
\alias{lmom_fast}
\title{Helper Function for L-moments Ratios}
\usage{
lmom_fast(model, params)
}
\arguments{
\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}
}
\value{
A numeric vector of length 4 containing:
\itemize{
\item \eqn{\lambda_1}: L-mean
\item \eqn{\lambda_2}: L-variance
\item \eqn{\tau_3}: L-skewness
\item \eqn{\tau_4}: L-kurtosis
}
}
\description{
A helper function used by \link{lmom_theoretical}.
This function does not validate parameters and is intended for internal use.
}
\note{
L-moment ratios for \code{NOR}/\code{LNO} and \code{PE3}/\code{LP3} are identical since
it L-moments for the \code{LNO}/\code{LP3} distributions are compared with the
sample L-moments of the logarithm of the data internally.
}
\examples{
lmom_fast("GLO", c(0, 1, 0))

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\link{lmom_theoretical}
}
