% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom-theoretical.R
\name{lmom_theoretical}
\alias{lmom_theoretical}
\alias{lmom_theoretical_gum}
\alias{lmom_theoretical_nor}
\alias{lmom_theoretical_gev}
\alias{lmom_theoretical_glo}
\alias{lmom_theoretical_gno}
\alias{lmom_theoretical_pe3}
\alias{lmom_theoretical_wei}
\title{Theoretical L-moments of Probability Distributions}
\usage{
lmom_theoretical_gum(params)

lmom_theoretical_nor(params)

lmom_theoretical_gev(params)

lmom_theoretical_glo(params)

lmom_theoretical_gno(params)

lmom_theoretical_pe3(params)

lmom_theoretical_wei(params)
}
\arguments{
\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}
}
\value{
A numeric vector of length 4 containing:
\itemize{
\item \eqn{\lambda_1}: L-mean
\item \eqn{\lambda_2}: L-variance
\item \eqn{\tau_3}: L-skewness
\item \eqn{\tau_4}: L-kurtosis
}
}
\description{
Computes the first four L-moments and L-moment ratios of seven different
probability distributions (\code{GUM}, \code{NOR}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, and \code{WEI})
given the parameters of the distribution.
}
\details{
The distributions \code{GUM}, \code{NOR}, \code{GEV}, \code{GLO}, and \code{WEI} have closed-form solutions
for the L-moments and L-moment ratios in terms of the parameters. The distributions
\code{GNO}, \code{PE3} use rational approximations of the L-moment ratios from Hosking (1997).
}
\examples{
lmom_theoretical_gev(c(0, 1, 0))

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\link{lmom_fast}
}
