% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom-theoretical.R
\name{lmr-functions}
\alias{lmr-functions}
\alias{lmrgum}
\alias{lmrnor}
\alias{lmrgev}
\alias{lmrglo}
\alias{lmrgno}
\alias{lmrpe3}
\alias{lmrwei}
\title{Theoretical L-moments of Probability Distributions}
\usage{
lmrgum(params)

lmrnor(params)

lmrgev(params)

lmrglo(params)

lmrgno(params)

lmrpe3(params)

lmrwei(params)
}
\arguments{
\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}
}
\value{
A numeric vector of length 4 containing:
\itemize{
\item \eqn{\lambda_1}: L-mean
\item \eqn{\lambda_2}: L-variance
\item \eqn{\tau_3}: L-skewness
\item \eqn{\tau_4}: L-kurtosis
}
}
\description{
Computes the first four L-moments and L-moment ratios of seven different
probability distributions (\code{GUM}, \code{NOR}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, and \code{WEI})
given the parameters of the distribution.
}
\details{
The distributions \code{GUM}, \code{NOR}, \code{GEV}, \code{GLO}, and \code{WEI} have closed-form solutions
for the L-moments and L-moment ratios in terms of the parameters. The distributions
\code{GNO} and \code{PE3} use rational approximations of the L-moment ratios from Hosking (1997).
}
\examples{
lmrgev(c(0, 1, 0))

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
