% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom-fast.R
\name{lmrxxx}
\alias{lmrxxx}
\title{Helper Function for L-moments Ratios}
\usage{
lmrxxx(model, params)
}
\arguments{
\item{model}{Character; three character distribution code. Must be one of:
\code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, or \code{WEI}.}

\item{params}{Numeric; a vector of parameters.
\itemize{
\item Numeric (2) if \code{model} is \code{GUM}, \code{NOR}, or \code{LNO}.
\item Numeric (3) if \code{model} is \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, or \code{WEI}.
}}
}
\value{
A numeric vector of length 4 containing:
\itemize{
\item \eqn{\lambda_1}: L-mean
\item \eqn{\lambda_2}: L-variance
\item \eqn{\tau_3}: L-skewness
\item \eqn{\tau_4}: L-kurtosis
}
}
\description{
A helper function used by \link{lmr-functions}.
}
\note{
L-moment ratios for \code{NOR}/\code{LNO} and \code{PE3}/\code{LP3} are identical
since it is assumed that the \code{LNO}/\code{LP3} L-moments will be compared
to the sample L-moments of the logarithm of the data.
}
\examples{
lmrxxx("GLO", c(0, 1, 0))

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
