% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik-fast.R
\name{loglik_fast}
\alias{loglik_fast}
\title{Log-Likelihood Helper Function}
\usage{
loglik_fast(data, model, params, years, trend)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\value{
Numeric scalar. The log-likelihood value.
}
\description{
A helper function used by \link{loglik_methods}.
This function does not validate parameters and is intended for internal use.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 0)
years <- seq(from = 1901, to = 2000)
trend <- list(location = FALSE, scale = FALSE)
loglik_fast(data, "GEV", params, years, trend)

}
\seealso{
\link{loglik_methods}
}
