% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik-methods.R
\name{loglik_methods}
\alias{loglik_methods}
\alias{loglik_gum}
\alias{loglik_nor}
\alias{loglik_lno}
\alias{loglik_gev}
\alias{loglik_glo}
\alias{loglik_gno}
\alias{loglik_pe3}
\alias{loglik_lp3}
\alias{loglik_wei}
\title{Log-Likelihood Functions for Probability Models}
\usage{
loglik_gum(data, params, years = NULL, trend = NULL)

loglik_nor(data, params, years = NULL, trend = NULL)

loglik_lno(data, params, years = NULL, trend = NULL)

loglik_gev(data, params, years = NULL, trend = NULL)

loglik_glo(data, params, years = NULL, trend = NULL)

loglik_gno(data, params, years = NULL, trend = NULL)

loglik_pe3(data, params, years = NULL, trend = NULL)

loglik_lp3(data, params, years = NULL, trend = NULL)

loglik_wei(data, params, years = NULL, trend = NULL)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as non-stationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\value{
Numeric scalar. The log-likelihood value.
}
\description{
Compute the log-likelihood value for stationary and non-stationary variants
of nine different distributions: \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"},
\code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.
}
\examples{
# Initialize data, params, years, and trend
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0)
years <- seq(from = 1901, to = 2000)
trend <- list(location = TRUE, scale = FALSE)

# Compute the log-likelihood
loglik_gno(data, params, years, trend)

}
\seealso{
\link{loglik_fast}
}
