% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-maximum-likelihood.R
\name{mle.estimation}
\alias{mle.estimation}
\title{Maximum Likelihood Parameter Estimation}
\usage{
mle.estimation(data, years, model, trend, prior = NULL)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{prior}{Numeric (2); optional vector of parameters \eqn{(p, q)} that specifies
the parameters of a Beta prior on \eqn{\kappa}. Only works with distribution \code{"GEV"}}
}
\value{
List; results of parameter estimation:
\itemize{
\item \code{params}: Numeric vector of estimated parameters.
\item \code{mll}: Maximum log‐likelihood value.
}
}
\description{
Estimates parameters of a probability distribution (\code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO},
\code{GNO}, \code{PE3}, \code{LP3}, or \code{WEI}) or non-stationary variant by maximizing the log‐likelihood.
Initial values are obtained through L‐moment parameter estimation, and optimization is
performed via \link[stats]{nlminb} with repeated perturbations if needed.
}
\details{
\enumerate{
\item Calls \link{pelxxx} on \code{data} to obtain initial parameter estimates.
\item Initializes trend parameters to zero if necessary.
\item For \code{WEI} models, sets the location parameter to zero to ensure support.
\item Defines an objective function as the negative of the \link{llvxxx} function.
\item Runs \link[stats]{nlminb} with box constraints. Attempts optimization
up to 100 times.
}
}
\note{
Although the more modern \link[stats]{optim} function is preferred over
\link[stats]{nlminb}, we use \link[stats]{nlminb} because it supports non-finite
values of the likelihood function.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
mle.estimation(data, years, "GNO100")

}
\seealso{
\link{llvxxx}, \link{gllxxx}, \link{pelxxx}, \link[stats]{nlminb}
}
