% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{model.assessment}
\alias{model.assessment}
\title{Evaluate Goodness-of-Fit for Fitted Flood Models}
\usage{
model.assessment(
  data,
  years,
  params,
  model,
  trend,
  uncertainty,
  pp.formula = "Weibull",
  alpha = 0.05
)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{params}{Numeric; a vector of fitted model parameters generated by
\link{pelxxx} or \link{mle.estimation}.}

\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{uncertainty}{List; estimated reutrn levels and confidence intervals
generated by \link{sb.uncertainty} or \link{rfpl.uncertainty}.}

\item{pp.formula}{Character (1); string specifying the plotting position formula.
One of \code{"Weibull"}, \code{"Blom"}, \code{"Cunnane"}, \code{"Gringorten"}, or \code{"Hazen"}
(default is "Weibull").}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}
}
\value{
List; model assessment metrics:
\itemize{
\item \code{estimates}: Quantile estimates for empirical return periods.
\item \code{R2}: Coefficient of determination from linear regression of estimates vs. data.
\item \code{RMSE}: Root mean squared error of quantile estimates.
\item \code{bias}: Mean bias of quantile estimates.
\item \code{AIC}: Akaike Information Criterion.
\item \code{BIC}: Bayesian Information Criterion.
\item \code{AIC_MLL}: Akaike Information Criterion, computed using the maximum log-likelihood.
\item \code{BIC_MLL}: Bayesian Information Criterion, computed using the maximum log-likelihood.
\item \code{AW}: Average width of the confidence interval(s).
\item \code{POC}: Percent of observations covered by the confidence interval(s).
\item \code{CWI}: Confidence width index, a metric that combines \code{AW} and \code{POC}.
}
}
\description{
Computes multiple performance metrics and diagnostic indicators to assess the quality of
a fitted flood frequency model This includes residual statistics, information criteria,
and coverage-based metrics using bootstrapped confidence intervals.
}
\examples{
# Initialize example data, years, and params
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
params <- c(100, 10)

# Perform uncertainty analysis
uncertainty <- sb.uncertainty(data, years, "NOR", "L-moments")

# Perform model assessment
model.assessment(data, years, "NOR", params, uncertainty)

}
\seealso{
\link{sb.uncertainty}, \link{rfpl.uncertainty},
\link[stats]{lm}, \link[stats]{approx}
}
