% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Evaluate Goodness-of-Fit for Fitted Flood Models}
\usage{
model_diagnostics(
  data,
  distribution,
  params,
  uncertainty,
  years = NULL,
  structure = NULL,
  alpha = 0.05,
  pp_formula = "Weibull"
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{distribution}{Character scalar. A three-character code indicating
a distribution family. Must be one of: \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}

\item{uncertainty}{List; estimated reutrn levels and confidence intervals
generated by \code{\link[=uncertainty_bootstrap]{uncertainty_bootstrap()}} or \code{\link[=uncertainty_rfpl]{uncertainty_rfpl()}}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{structure}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{pp_formula}{Character (1); string specifying the plotting position formula.
Must be one of \code{"Weibull"}, \code{"Blom"}, \code{"Cunnane"}, \code{"Gringorten"}, or \code{"Hazen"}}
}
\value{
List containing model assessment metrics:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{estimates}: Quantile estimates for empirical return periods.
\item \code{R2}: Coefficient of determination from linear regression of estimates vs. data.
\item \code{RMSE}: Root mean squared error of quantile estimates.
\item \code{bias}: Mean bias of quantile estimates.
\item \code{AIC}: Akaike Information Criterion.
\item \code{BIC}: Bayesian Information Criterion.
\item \code{AIC_MLL}: Akaike Information Criterion, computed using MLE.
\item \code{BIC_MLL}: Bayesian Information Criterion, computed using MLE.
\item \code{AW}: Average width of the confidence interval(s).
\item \code{POC}: Percent of observations covered by the confidence interval(s).
\item \code{CWI}: Confidence width index, a metric that combines \code{AW} and \code{POC}.
}
}
\description{
Computes multiple performance metrics and diagnostic indicators to assess the
quality of a fitted flood frequency model. This includes accuracy (residual
statistics), fitting efficiency (information criteria), and uncertainty
(coverage-based metrics using confidence intervals).
}
\examples{
# Initialize example data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(100, 10)

# Perform uncertainty analysis
uncertainty <- uncertainty_bootstrap(data, "NOR", "L-moments")

# Evaluate model diagnostics
model_diagnostics(data, "NOR", params, uncertainty)

}
\seealso{
\code{\link[=uncertainty_bootstrap]{uncertainty_bootstrap()}}, \code{\link[=uncertainty_rfpl]{uncertainty_rfpl()}}, \code{\link[=fit_maximum_likelihood]{fit_maximum_likelihood()}}
}
