% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-eda.R
\name{module_eda}
\alias{module_eda}
\title{Conduct Exploratory Data Analysis}
\usage{
module_eda(data, years, splits = NULL, automatic = FALSE, ...)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{splits}{An integer vector of years at which to split the data (default is \code{NULL}).
A split point is the \emph{first} year in a subperiod.}

\item{automatic}{If \code{TRUE}, the data is split automatically using the results of the Pettitt
and MKS tests (default is \code{FALSE}). This argument is ignored if \code{splits} is not \code{NULL}.}

\item{...}{Additional arguments to be passed to the statistical tests. See details.}
}
\value{
\code{recommendations}: A list of recommended split points and nonstationary structures:
\itemize{
\item \code{splits}: The change point(s) identified by the change point detection test with the
the lowest statistically significant p-value, or an empty vector if no such test exists.
\item \code{structures}: A list of nonstationary structure objects for each subperiod. Each structure
is a list with boolean items \code{location} and \code{scale} which represent a linear trend in the
in the mean or variability of the data respectively.
}

\code{blocks}: A list of grouped statistical tests. Each block is a list containing:
\itemize{
\item \code{name}: Either \code{"Change Points"} or \code{"Trend Detection"}.
\item \code{start}: The first year of the subperiod.
\item \code{end}: The last year of the subperiod.
\item Additional lists corresponding to the individual statistical tests within the block.
}
}
\description{
Identifies change points in annual maximum series (AMS) data. Allows the user to split
the dataset at one or more years. Finally, performs a collection of statistical tests
for identifying nonstationarity in the mean/variability of each subperiod.
}
\details{
The behaviour of the \code{splits} and \code{automatic} arguments can be confusing:
\itemize{
\item If \code{splits = NULL} and \code{automatic = TRUE}, the data will be split automatically based on
the change point detection test with the lowest statistically significant p-value.
If neither test yielded a statistically significant p-value, the data will not be split.
\item If \code{splits = NULL} and \code{automatic = FALSE}, the user will be prompted to select the split
points. If R is not running in interactive mode, the function will return the results of
change point detection along with an error message.
\item If splits is not \code{NULL}, it will be used to split the data. To avoid splitting the data,
use \code{splits = integer(0)}. Note that \code{integer(0)} is NOT the same as \code{NULL}.
}

In addition to the required arguments, \code{module_eda} also accepts:
\itemize{
\item \code{alpha}: The numeric significance level for all statistical tests (default is 0.05).
\item \code{bbmk_samples}: The number of samples used in the Block-Bootstrap Mann-Kendall (BBMK)
test (default is 10000). Must be an integer.
}
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
module_eda(data, years, automatic = TRUE)

}
\seealso{
\code{\link[=eda_pettitt_test]{eda_pettitt_test()}}, \code{\link[=eda_mks_test]{eda_mks_test()}}, \code{\link[=eda_mk_test]{eda_mk_test()}}, \code{\link[=eda_spearman_test]{eda_spearman_test()}}
\code{\link[=eda_bbmk_test]{eda_bbmk_test()}}, \code{\link[=eda_pp_test]{eda_pp_test()}}, \code{\link[=eda_kpss_test]{eda_kpss_test()}}, \code{\link[=eda_sens_trend]{eda_sens_trend()}},
\code{\link[=eda_runs_test]{eda_runs_test()}}, \code{\link[=eda_white_test]{eda_white_test()}}
}
