% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-kappa.R
\name{mu.sigma}
\alias{mu.sigma}
\title{Compute Location and Scale of Kappa Distribution}
\usage{
mu.sigma(l1, l2, k, h)
}
\description{
Compute Location and Scale of Kappa Distribution
}
\keyword{internal}
