% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmoments-kappa.R
\name{mu_sigma}
\alias{mu_sigma}
\title{Compute Location and Scale of Kappa Distribution}
\usage{
mu_sigma(l1, l2, k, h)
}
\description{
Compute Location and Scale of Kappa Distribution
}
\keyword{internal}
