% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-mw-variance.R
\name{mw.variance}
\alias{mw.variance}
\title{Estimate Variance for Annual Maximum Streamflow Data}
\usage{
mw.variance(data, years, size = 10, step = 5)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{size}{Integer (1); the number of consecutive indices in each moving
window. Must be a positive integer less than or equal to \code{length(data)}
(default is 10). If \code{length(data) < size}, an error is raised.}

\item{step}{Integer (1); the offset (in indices) between successive moving
windows. Must be a positive integer (default is 5).}
}
\value{
A list with two entries:
\itemize{
\item \code{years}: Numeric; the mean year within each window.
\item \code{std}: Numeric; the standard deviation of the data within each window.
}
}
\description{
This function estimates the standard deviation of a vector of annual maximum
streamflow (AMS) data using a moving window algorithm, returning a list that
pairs each window’s mean year with its computed standard deviation. The
parameters \code{size} and \code{step} parameters control the behaviour of the window.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
mw.variance(data, years)

}
