% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-alpha}
\alias{param-alpha}
\title{Parameter 'alpha'}
\arguments{
\item{alpha}{Numeric scalar in \link{0.01, 0.1}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}
}
\description{
Parameter 'alpha'
}
