% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-distribution}
\alias{param-distribution}
\title{Parameter 'distribution'}
\arguments{
\item{distribution}{A three-character code indicating the distribution family.
Must be \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"},
or \code{"WEI"}.}
}
\description{
Parameter 'distribution'
}
\keyword{internal}
