% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-generate-report}
\alias{param-generate-report}
\title{Parameter 'generate_report'}
\arguments{
\item{generate_report}{If \code{TRUE} (default), generate a report.}
}
\description{
Parameter 'generate_report'
}
\keyword{internal}
