% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-method}
\alias{param-method}
\title{Parameter 'method'}
\arguments{
\item{method}{Character scalar specifying the estimation method.
Must be \code{"L-moments"}, \code{"MLE"}, or \code{"GMLE"}.}
}
\description{
Parameter 'method'
}
\keyword{internal}
