% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-slice}
\alias{param-ns-slice}
\title{Parameter 'ns_slice'}
\arguments{
\item{ns_slice}{For NS-FFA only: Numeric scalar specifying the year at which to
evaluate  the quantiles of a nonstationary probability distribution. \code{ns_slice}
does not have to be an element of the \code{ns_years} argument.}
}
\description{
Parameter 'ns_slice'
}
\keyword{internal}
