% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-slices}
\alias{param-ns-slices}
\title{Parameter 'ns_slices'}
\arguments{
\item{ns_slices}{NS-FFA only. Numeric vector specifying the years at which to evaluate
the return levels confidence intervals of a nonstationary probability distribution.
\code{ns_slices} do not have to be elements of the \code{ns_years} argument.}
}
\description{
Parameter 'ns_slices'
}
\keyword{internal}
