% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-structure}
\alias{param-ns-structure}
\title{Parameter 'ns_structure'}
\arguments{
\item{ns_structure}{For NS-FFA only: Named list indicating which distribution
parameters are modeled as nonstationary. Must contain two logical scalars:
\itemize{
\item \code{location}: If \code{TRUE}, the location parameter has a linear temporal trend.
\item \code{scale}: If \code{TRUE}, the scale parameter has a linear temporal trend.
}}
}
\description{
Parameter 'ns_structure'
}
\keyword{internal}
