% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-structures}
\alias{param-ns-structures}
\title{Parameter 'ns_structures'}
\arguments{
\item{ns_structures}{For S-FFA, set to \code{NULL} (default) to use a stationary model
for all homogeneous subperiods. For NS-FFA, provide a list of \code{length(ns_splits) + 1}
sublists specifying the nonstationary model structure for each homogeneous subperiod.
Each sublist must contain logical elements \code{location} and \code{scale}, indicating
monotonic trends in the mean and variability, respectively.}
}
\description{
Parameter 'ns_structures'
}
\keyword{internal}
