% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-years}
\alias{param-ns-years}
\title{Parameter 'ns_years'}
\arguments{
\item{ns_years}{NS-FFA only. Numeric vector of observation years corresponding
to \code{data}. Must be the same length as \code{data} and strictly increasing.}
}
\description{
Parameter 'ns_years'
}
\keyword{internal}
