% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-params}
\alias{param-params}
\title{Parameter 'params'}
\arguments{
\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}
}
\description{
Parameter 'params'
}
\keyword{internal}
