% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{param-periods}
\alias{param-periods}
\title{Parameter 'periods'}
\arguments{
\item{periods}{Numeric vector used to set the return periods for FFA.
All entries must be greater than or equal to 1.}
}
\description{
Parameter 'periods'
}
\keyword{internal}
