% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{param-quiet}
\alias{param-quiet}
\title{Parameter 'quiet'}
\arguments{
\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\description{
Parameter 'quiet'
}
\keyword{internal}
