% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-report-formats}
\alias{param-report-formats}
\title{Parameter 'report_formats'}
\arguments{
\item{report_formats}{A character vector specifying the output format for the
report. Supported values are \code{"md"}, \code{"pdf"}, \code{"html"}, and \code{"json"}.}
}
\description{
Parameter 'report_formats'
}
\keyword{internal}
