% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-report-path}
\alias{param-report-path}
\title{Parameter 'report_path'}
\arguments{
\item{report_path}{A character scalar, the file path for the generated report.
If \code{NULL} (default), the report will be saved to a new temporary directory.}
}
\description{
Parameter 'report_path'
}
\keyword{internal}
