% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{param-slice}
\alias{param-slice}
\title{Parameter 'slice'}
\arguments{
\item{slice}{Numeric scalar specifying the year at which to evaluate the
quantiles of a nonstationary probability distribution.  The slice does not
have to be an element of the \code{years} argument. Note that if \code{trend$location}
and \code{trend$scale} are both \code{FALSE}, this argument will have  no effect the
output of the function.}
}
\description{
Parameter 'slice'
}
\keyword{internal}
