% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{param-slices}
\alias{param-slices}
\title{Parameter 'slices'}
\arguments{
\item{slices}{Numeric vector specifying the years at which to evaluate the
return levels confidence intervals of a nonstationary probability distribution.
The slices do not have to be elements of the \code{years} argument.}
}
\description{
Parameter 'slices'
}
\keyword{internal}
