% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-structure}
\alias{param-structure}
\title{Parameter 'structure'}
\arguments{
\item{structure}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain two logical scalars:
\itemize{
\item \code{location}: If \code{TRUE}, the location parameter has a linear temporal trend.
\item \code{scale}: If \code{TRUE}, the scale parameter has a linear temporal trend.
}}
}
\description{
Parameter 'structure'
}
\keyword{internal}
