% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-tolerance}
\alias{param-tolerance}
\title{Parameter 'tolerance'}
\arguments{
\item{tolerance}{The log-likelihood tolerance for Regula-Falsi convergence
(default is 0.01).}
}
\description{
Parameter 'tolerance'
}
\keyword{internal}
