% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-trend}
\alias{param-trend}
\title{Parameter 'trend'}
\arguments{
\item{trend}{Named list indicating which distribution parameters are
modeled as non-stationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\description{
Parameter 'trend'
}
\keyword{internal}
