% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-years}
\alias{param-years}
\title{Parameter 'years'}
\arguments{
\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}
}
\description{
Parameter 'years'
}
