% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pel-functions.R
\name{pel-functions}
\alias{pel-functions}
\alias{pelgum}
\alias{pelnor}
\alias{pellno}
\alias{pelgev}
\alias{pelglo}
\alias{pelgno}
\alias{pelpe3}
\alias{pellp3}
\alias{pelwei}
\title{Parameter Estimation with L-Moments}
\usage{
pelgum(data)

pelnor(data)

pellno(data)

pelgev(data)

pelglo(data)

pelgno(data)

pelpe3(data)

pellp3(data)

pelwei(data)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}
}
\value{
Numeric; a vector of parameters.
}
\description{
Estimate the parameters of one of nine different distributions (\code{GUM}, \code{NOR},
\code{LNO}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, and \code{WEI}) using the method of L-moments.
}
\details{
First, the sample L-moments of the data are computed using the \link{lmom.sample}
method. Then formulas from Hosking (1997) are used to compute the parameters from the
L-moments. Distributions \code{GNO}, \code{PE3}, and \code{LP3} use a rational approximation to compute
the parameters.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
pellp3(data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
