% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-kappa.R
\name{pelkap}
\alias{pelkap}
\title{L-Moments Parameter Estimation for the Kappa Distribution}
\usage{
pelkap(data)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}
}
\value{
Numeric (4); a vector of parameters.
}
\description{
This functions estimates the parameters of the four-parameter Kappa distribution using
the method of L-moments. Since there is no known closed form solution for the parameters
in terms of the L-moments, the parameters are computed numerically using Newton-Raphson
iteration.
}
\details{
First, the sample L-moments of the data are computed using the \link{lmom.sample}
method. Then, the \link[stats]{optim} function is used to determine the
parameters by minimizing the \code{sumquad.tau3tau4} helper function. The implementation of
this routine is based on the deprecated \code{homtest} package.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
pelkap(data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
