% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pel-xxx.R
\name{pelxxx}
\alias{pelxxx}
\title{Helper Function for L-moments Parameter Estimation}
\usage{
pelxxx(model, data)
}
\arguments{
\item{model}{Character; \code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO}, \code{GNO},
\code{PE3}, \code{LP3}, or \code{WEI}.}

\item{data}{Numeric; a vector of annual maximum streamflow data.}
}
\value{
Numeric; a vector of parameters.
\itemize{
\item Numeric (2) if \code{model} is \code{GUM}, \code{NOR}, or \code{LNO}.
\item Numeric (3) if \code{model} is \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, or \code{WEI}.
}
}
\description{
A helper function used by \link{pel-functions}.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
pelxxx("PE3", data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
