% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-pettitt-test.R
\name{pettitt_test}
\alias{pettitt_test}
\title{Pettitt Test for Abrupt Changes in the Mean of a Time Series}
\usage{
pettitt_test(ams, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{ams}{List; annual maximum streamflow (AMS) data:
\itemize{
\item \code{data} Numeric; a vector of AMS values or moving-window variances.
\item \code{year} Numeric; a vector of years with the same length as \code{data}.
\item \code{variance} Logical (1); if \code{TRUE}, \code{data} contains AMS variances.
}}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}

\item{quiet}{Logical (1); if FALSE, prints a summary of results (default is TRUE).}
}
\value{
List; the results of the Pettitt test, including:
\itemize{
\item \code{u_t} Numeric; vector of absolute U-statistics for all time indices.
\item \code{k_statistic} Numeric (1); maximum absolute U-statistic.
\item \code{k_critical} Numeric (1);  critical K-statistic value for given \code{alpha}.
\item \code{p_value} Numeric (1); Approximate p-value for the test.
\item \code{change_index} Integer (1); index of the detected change point (0 if none).
\item \code{change_year} Integer (1); year of the detected change point (0 if none).
\item \code{reject} Logical (1); if TRUE, the null hypothesis was rejected.
\item \code{msg} Character (1); formatted summary message describing the test result.
}
}
\description{
Pettitt Test for Abrupt Changes in the Mean of a Time Series
}
\details{
The Pettitt test is a rank-based non-parametric test that evaluates the
hypothesis of a change point in the mean of a time series. It computes the
maximum of the absolute value of the U-statistic over all possible split
points. The p-value is approximated using an asymptotic formula.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
ams <- list(data = data, year = year, variance = FALSE)
pettitt.test(data, years)
}
\seealso{
\link{pettitt.plot}
}
