% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-model-diagnostics.R
\name{plot_model_diagnostics}
\alias{plot_model_diagnostics}
\title{Plot Model Assessment Results}
\usage{
plot_model_diagnostics(results, ...)
}
\arguments{
\item{results}{List; model assessment results generated by \code{\link[=model_diagnostics]{model_diagnostics()}}.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item A black line representing a model with no deviation from the emprical quantiles.
\item Red points denoting the estimated quantiles against the empirical quantiles.
}
}
\description{
Creates a quantile–quantile plot comparing observed annual maximum series
data to quantile estimates from a fitted parametric model. The 1:1 line is drawn in
black and the parametric model estimates are plotted as semi‐transparent red points.
}
\examples{
# Initialize example data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(100, 10)

# Perform uncertainty analysis
uncertainty <- uncertainty_bootstrap(data, "NOR", "L-moments")

# Evaluate model diagnostics
results <- model_diagnostics(data, "NOR", params, uncertainty)

# Generate a model assessment plot
plot_model_diagnostics(results)

}
