% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-nsffa.R
\name{plot_nsffa}
\alias{plot_nsffa}
\title{Plot Return Levels and Confidence Intervals for NS-FFA}
\usage{
plot_nsffa(results, ...)
}
\arguments{
\item{results}{A list of estimated return levels and confidence intervals
generated by \code{\link[=uncertainty_bootstrap]{uncertainty_bootstrap()}} or \code{\link[=uncertainty_rfpl]{uncertainty_rfpl()}}.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot with one line and ribbon per slice.
}
\description{
Generates a plot with effective return periods on the x-axis and effective return
levels (annual maxima magnitudes) on the y-axis for up to 5 time slices. Each slice
is displayed in a distinct color. Confidence bounds are shown as semi-transparent
ribbons, and the point estimates  are overlaid as solid lines. Return periods are
depicted on a logarithmic scale.
}
\examples{

data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)

# Run the uncertainty bootstrap at slices 1920, 1960, 2000
results <- uncertainty_bootstrap(
	   data,
   "GEV",
   "L-moments",
   years = years,
   structure = list(location = TRUE, scale = FALSE),
   slices = c(1920, 1960, 2000),
	   samples = 1000L
)

# Generate the plot
plot_nsffa(results)

}
