% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-nsffa-assessment.R
\name{plot_nsffa_assessment}
\alias{plot_nsffa_assessment}
\title{Plot Model Assessment for NS-FFA}
\usage{
plot_nsffa_assessment(results, ...)
}
\arguments{
\item{results}{List; model assessment results generated by \code{\link[=model_assessment]{model_assessment()}}.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item A black line representing a model with no deviation from the empirical quantiles.
\item Red points denoting the estimated quantiles against the empirical quantiles.
}
}
\description{
Creates a normalized and detrended quantile–quantile plot (a worm plot) comparing
empirical annual maximum series data to quantile estimates from a fitted, parametric,
and nonstationary model. Confidence intervals are also provided. The worm plot can
be interpreted using the following heuristics:
\itemize{
\item If the worm (red points) is consistently above/below the origin, the model mean
is too large/small respectively.
\item If the worm has a clear positive/negative slope, the model variance is too
large/small respectively.
\item If the worm has a U-shape or inverted U-shape, the model skewness is too
small/large respectively.
}
}
\examples{
# Initialize example data and params
data <- rnorm(n = 100, mean = 100, sd = 10) + seq(from = 1, to = 100)
years <- seq(from = 1901, to = 2000)
structure <- list(location = TRUE, scale = FALSE)

# Evaluate model diagnostics
results <- model_assessment(data, "NOR", "MLE", NULL, years, structure)

# Generate a model assessment plot
plot_nsffa_assessment(results)

}
