% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sens-trend.R
\name{plot_sens_trend}
\alias{plot_sens_trend}
\title{Plot Sen’s Trend Estimator}
\usage{
plot_sens_trend(
  data,
  years,
  mean_trend = NULL,
  variability_trend = NULL,
  show_line = TRUE,
  ...
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{mean_trend}{Trend in the mean estimated by by \code{\link[=eda_sens_trend]{eda_sens_trend()}}.}

\item{variability_trend}{Trend in variability estimated by \code{\link[=eda_sens_trend]{eda_sens_trend()}}.}

\item{show_line}{If \code{TRUE} (default), draw a fitted line through the data.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item Gray points for each year’s annual maximum series value.
\item Optional gray line connecting the data if \code{show_line = TRUE}.
\item A solid black line representing a constant mean, if \code{mean_trend == NULL}.
\item A solid blue line representing a trend in the mean, if \code{mean_trend != NULL}.
\item A dashed black line representing constant variability, if \code{variability_trend == NULL}.
\item A dashed blue line representing a trend in variability, if \code{variability_trend != NULL}.
\item The equation for the trend in the mean, written in the form \eqn{mx + b}.
}
}
\description{
Produces a scatterplot of the annual maximum series data or its variance
against time, optionally overlaid with Sen’s trend estimator of the mean and/or
variability.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
mean_trend <- eda_sens_trend(data, years)
plot_sens_trend(data, years, meant_trend = mean_trend)

}
\seealso{
\code{\link[=eda_sens_trend]{eda_sens_trend()}}
}
