% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sffa.R
\name{plot_sffa}
\alias{plot_sffa}
\title{Plot Return Levels and Confidence Intervals for S-FFA}
\usage{
plot_sffa(results, ...)
}
\arguments{
\item{results}{A list of estimated return levels and confidence intervals
generated by \code{\link[=uncertainty_bootstrap]{uncertainty_bootstrap()}} or \code{\link[=uncertainty_rfpl]{uncertainty_rfpl()}}.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot showing:
\itemize{
\item A semi-transparent gray ribbon between \code{results$ci_lower} and \code{results$ci_upper}.
\item A solid black line for the point estimates.
}
}
\description{
Generates a plot with return periods on the x-axis and return levels (annual
maxima magnitudes) on the y-axis for stationary flood frequency analysis.
The confidence bound is shown as a semi-transparent ribbon, and the point estimates
are overlaid as a solid line. Return periods are shown on a logarithmic scale.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
results <- uncertainty_bootstrap(data, "WEI", "L-moments")
plot_sffa(results)

}
