% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-methods.R
\name{qnt-functions}
\alias{qnt-functions}
\alias{qntgum}
\alias{qntgum10}
\alias{qntgum11}
\alias{qntnor}
\alias{qntnor10}
\alias{qntnor11}
\alias{qntlno}
\alias{qntlno10}
\alias{qntlno11}
\alias{qntgev}
\alias{qntgev100}
\alias{qntgev110}
\alias{qntglo}
\alias{qntglo100}
\alias{qntglo110}
\alias{qntgno}
\alias{qntgno100}
\alias{qntgno110}
\alias{qntpe3}
\alias{qntpe3100}
\alias{qntpe3110}
\alias{qntlp3}
\alias{qntlp3100}
\alias{qntlp3110}
\alias{qntwei}
\alias{qntwei100}
\alias{qntwei110}
\alias{qntkap}
\title{Quantile Functions for Probability Models}
\usage{
qntgum(p, params, years = NULL)

qntgum10(p, params, years)

qntgum11(p, params, years)

qntnor(p, params, years = NULL)

qntnor10(p, params, years)

qntnor11(p, params, years)

qntlno(p, params, years = NULL)

qntlno10(p, params, years)

qntlno11(p, params, years)

qntgev(p, params, years = NULL)

qntgev100(p, params, years)

qntgev110(p, params, years)

qntglo(p, params, years = NULL)

qntglo100(p, params, years)

qntglo110(p, params, years)

qntgno(p, params, years = NULL)

qntgno100(p, params, years)

qntgno110(p, params, years)

qntpe3(p, params, years = NULL)

qntpe3100(p, params, years)

qntpe3110(p, params, years)

qntlp3(p, params, years = NULL)

qntlp3100(p, params, years)

qntlp3110(p, params, years)

qntwei(p, params, years = NULL)

qntwei100(p, params, years)

qntwei110(p, params, years)

qntkap(p, params, years = NULL)
}
\arguments{
\item{p}{Numeric; a vector of probabilities between 0 and 1.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.
Required for non-stationary models, which end in \code{10}, \code{11}, \code{100}, or \code{110}.}
}
\value{
If \code{p} or \code{years} is a scalar, returns a numeric vector. Otherwise, returns a matrix.
}
\description{
Compute the quantiles for stationary and non-stationary variants
of nine different distributions (\code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO},
\code{GNO}, \code{PE3}, \code{LP3}, and \code{WEI}).
}
\details{
The quantile function is the inverse of the cumulative distribution function.
For the two-parameter distributions (\code{GUM}, \code{NOR}, \code{LNO}), there are three
different \code{qnt} functions:
\itemize{
\item \code{qnt...()}: Stationary location and scale, 2 parameters.
\item \code{qnt...10()}: Time-varying location, stationary scale, 3 parameters.
\item \code{qnt...11()}: Time-varying location and scale, 4 parameters.
}

For three-parameter distributions (\code{GEV}, \code{GLO}, \code{GNO}, \code{PE3}, \code{LP3}, \code{WEI}),
there are also three \code{qnt} functions:
\itemize{
\item \code{qnt...()}: Stationary location and scale, 3 parameters.
\item \code{qnt...100()}: Time-varying location, stationary scale, 4 parameters.
\item \code{qnt...110()}: Time-varying location and scale, 5 parameters.
}
}
\note{
The \code{qnt...}, functions perform extensive parameter validation, which can be slow.
If you plan to make many calls to these methods, it is recommended to use
the \link{qntxxx} helper function instead.
}
\examples{
# Initialize p, years, and params
p <- runif(n = 10)
years <- seq(from = 1901, to = 2000)
params <- c(0, 1, 1, 1, 0)

# Compute the quantiles
qntwei110(p, params, years)

}
\seealso{
\link{qntxxx}
}
