% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnt-xxx.R
\name{qntxxx}
\alias{qntxxx}
\title{Helper Function for Quantile Functions}
\usage{
qntxxx(name, signature, p, params, covariate = 0)
}
\arguments{
\item{name}{Character (1); the name of the probability distribution.}

\item{signature}{Character (1); the non-stationary signature (\code{NULL}, \code{"10"}, or \code{"11"}).}

\item{p}{Numeric; a vector of probabilities. Must be between 0 and 1.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{covariate}{Numeric; a vector with the same length as \code{data}.
Required if \code{signature} is \code{"10"} or \code{"11"}.}
}
\value{
If \code{p} or \code{years} is a scalar, returns a numeric vector. Otherwise, returns a matrix.
}
\description{
A helper function used by \link{qnt-functions}.
}
\examples{
# Initialize p and params
p <- runif(n = 10)
params <- c(0, 1, 0)

# Compute the log-likelihood
qntxxx("GEV", NULL, p, params)

}
\seealso{
\link{qnt-functions}
}
