% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-fast.R
\name{quantile_fast}
\alias{quantile_fast}
\title{Helper Function for Quantile Functions}
\usage{
quantile_fast(p, distribution, params, slice, structure)
}
\arguments{
\item{p}{Numeric vector of probabilities between 0 and 1 with no missing values.}

\item{distribution}{A three-character code indicating the distribution family.
Must be \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"},
or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}
}
\value{
A numeric vector of quantiles with the same length as \code{p}.
}
\description{
A helper function used by \code{\link[=quantile_xxx]{quantile_xxx()}}.
This function does not validate parameters and is designed for use in other methods.
}
\examples{
# Initialize p, params, and structure
p <- runif(n = 10)
params <- c(0, 1, 1, 0)
structure <- list(location = FALSE, scale = TRUE)

# Compute the log-likelihood in the year 2000
quantile_fast(p, "GEV", params, 2000, structure)

}
\seealso{
\code{\link[=quantile_xxx]{quantile_xxx()}}
}
