% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-methods.R
\name{quantile_methods}
\alias{quantile_methods}
\alias{quantile_gum}
\alias{quantile_nor}
\alias{quantile_lno}
\alias{quantile_gev}
\alias{quantile_glo}
\alias{quantile_gno}
\alias{quantile_pe3}
\alias{quantile_lp3}
\alias{quantile_wei}
\alias{quantile_kap}
\title{Quantile Functions for Probability Models}
\usage{
quantile_gum(p, params, slice = 1900, trend = NULL)

quantile_nor(p, params, slice = 1900, trend = NULL)

quantile_lno(p, params, slice = 1900, trend = NULL)

quantile_gev(p, params, slice = 1900, trend = NULL)

quantile_glo(p, params, slice = 1900, trend = NULL)

quantile_gno(p, params, slice = 1900, trend = NULL)

quantile_pe3(p, params, slice = 1900, trend = NULL)

quantile_lp3(p, params, slice = 1900, trend = NULL)

quantile_wei(p, params, slice = 1900, trend = NULL)

quantile_kap(p, params, slice = 1900, trend = NULL)
}
\arguments{
\item{p}{Numeric vector of probabilities between 0 and 1 with no missing values.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{model} and \code{trend}.}

\item{slice}{Numeric scalar specifying the year at which to evaluate the
quantiles of a non-stationary probability distribution.  The slice does not
have to be an element of the \code{years} argument. Note that if \code{trend$location}
and \code{trend$scale} are both \code{FALSE}, this argument will have  no effect the
output of the function.}

\item{trend}{Named list indicating which distribution parameters are
modeled as non-stationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}
}
\value{
A numeric vector of quantiles with the same length as \code{p}.
}
\description{
Compute the quantiles for stationary and non-stationary variants  of nine
different distributions: \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"},
\code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.
}
\examples{
# Initialize p and params
p <- runif(n = 10)
params <- c(0, 1, 0)

# Compute the quantiles
quantile_wei(p, params)

}
\seealso{
\link{quantile_fast}
}
