% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-xxx.R
\name{quantile_xxx}
\alias{quantile_xxx}
\alias{quantile_gum}
\alias{quantile_nor}
\alias{quantile_lno}
\alias{quantile_gev}
\alias{quantile_glo}
\alias{quantile_gno}
\alias{quantile_pe3}
\alias{quantile_lp3}
\alias{quantile_wei}
\alias{quantile_kap}
\title{Quantile Functions for Probability Models}
\usage{
quantile_gum(p, params, slice = 1900, structure = NULL)

quantile_nor(p, params, slice = 1900, structure = NULL)

quantile_lno(p, params, slice = 1900, structure = NULL)

quantile_gev(p, params, slice = 1900, structure = NULL)

quantile_glo(p, params, slice = 1900, structure = NULL)

quantile_gno(p, params, slice = 1900, structure = NULL)

quantile_pe3(p, params, slice = 1900, structure = NULL)

quantile_lp3(p, params, slice = 1900, structure = NULL)

quantile_wei(p, params, slice = 1900, structure = NULL)

quantile_kap(p, params, slice = 1900, structure = NULL)
}
\arguments{
\item{p}{Numeric vector of probabilities between 0 and 1 with no missing values.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}

\item{slice}{Numeric scalar specifying the year at which to evaluate the
quantiles of a nonstationary probability distribution. The slice does not
have to be an element of the \code{years} argument. Note that if \code{structure$location}
and \code{structure$scale} are both \code{FALSE}, this argument will have no effect the
output of the function.}

\item{structure}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain two logical scalars:
\itemize{
\item \code{location}: If \code{TRUE}, the location parameter has a linear temporal trend.
\item \code{scale}: If \code{TRUE}, the scale parameter has a linear temporal trend.
}}
}
\value{
A numeric vector of quantiles with the same length as \code{p}.
}
\description{
Compute the quantiles for stationary and nonstationary variants of nine
different distributions: \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"},
\code{"PE3"}, \code{"LP3"}, or \code{"WEI"}. In total, these methods compute the quantiles
for 36 different probability models.
}
\examples{
# Initialize p and params
p <- runif(n = 10)
params <- c(0, 1, 0)

# Compute the quantiles
quantile_wei(p, params)

}
\seealso{
\code{\link[=quantile_fast]{quantile_fast()}}
}
