% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs-plot.R
\name{runs.plot}
\alias{runs.plot}
\title{Plot Runs Test Results}
\usage{
runs.plot(years, residuals, results, name, show_trend = TRUE)
}
\arguments{
\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{residuals}{Numeric; A vector of residuals produced by \link{sens.trend}.}

\item{results}{List; runs test results generated by \link{runs.test}.}

\item{name}{Character; either \code{"sens-variance"} or \code{"sens-mean"}.}

\item{show_trend}{Logical; if TRUE (default), draw a fitted line through the AMS data.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item Black points for the residual at each year.
\item A red dashed horizontal line at \eqn{y = 0}.
\item A text annotation “Runs p-value: X.XXX” in the plot area.
}
}
\description{
Generates a residual plot of Sen’s estimator applied to annual maximum streamflow
(AMS) data with a horizontal dashed line at zero and an annotation of the Runs
test p-value.
}
\examples{
# Initialize data and years
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)

# Generate runs test plot 
residuals <- sens.trend(data, years)$residuals
results <- runs.test(residuals)
runs.plot(years, residuals, results, "sens-mean")

}
\seealso{
\link{runs.test}
}
