% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens-plot.R
\name{sens.plot}
\alias{sens.plot}
\title{Plot Sen’s Trend Estimator}
\usage{
sens.plot(data, years, results, name, show_trend = TRUE)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{results}{List; runs test results generated by \link{sens.trend}.}

\item{name}{Character; either \code{"sens-variance"} or \code{"sens-mean"}.}

\item{show_trend}{Logical; if TRUE (default), draw a fitted line through the AMS data.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item Black points for each year’s AMS (or variance) value.
\item Optional black line connecting the AMS data if \code{show_trend = TRUE}.
\item Blue line representing Sen’s trend estimator.
\item A text annotation displaying the fitted equation \eqn{y = mx + b}.
}
}
\description{
Produces a scatter plot of the annual maximum streamflow (AMS) data or its variance
against time, overlaid with Sen’s trend estimator and an annotation of the fitted
equation.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
results <- sens.trend(data, years)
sens.plot(data, years, results, "sens-mean")

}
\seealso{
\link{sens.trend}
}
