% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-kappa.R
\name{sumquad_tau3tau4}
\alias{sumquad_tau3tau4}
\title{Compute L-moment Distance for Kappa Distribution}
\usage{
sumquad_tau3tau4(k.h, t3.t4)
}
\description{
Compute L-moment Distance for Kappa Distribution
}
\keyword{internal}
