% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-bootstrap.R
\name{uncertainty_bootstrap}
\alias{uncertainty_bootstrap}
\title{Parametric Bootstrap Confidence Intervals for Flood Quantile Estimates}
\usage{
uncertainty_bootstrap(
  data,
  model,
  method,
  prior = NULL,
  years = NULL,
  trend = NULL,
  slices = 1900,
  alpha = 0.05,
  samples = 10000L,
  periods = c(2, 5, 10, 20, 50, 100)
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{method}{Character scalar specifying the estimation method.
Must be \code{"L-moments"}, \code{"MLE"}, or \code{"GMLE"}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{model = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}

\item{slices}{Numeric vector specifying the years at which to evaluate the
return levels confidence intervals of a nonstationary probability distribution.
The slices do not have to be elements of the \code{years} argument.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{samples}{Integer scalar. The number of bootstrap samples. Default is 10000.}

\item{periods}{Numeric vector used to set the return periods for FFA.
All entries must be greater than or equal to 1.}
}
\value{
A list of lists containing the return levels and confidence
intervals for each slice. Each sub-list contains:
\itemize{
\item \code{method}: "Bootstrap"
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{t}: Vector of return periods; \code{c(2, 5, 10, 20, 50, 100)}.
\item \code{slice}: The value of \code{slice} argument.
\item \code{trend}: The value of \code{trend} argument.
}
}
\description{
Computes estimates and confidence intervals for return levels at return periods
2, 5, 10, 20, 50, and 100 years using the parametric bootstrap method. This function
supports a variety of probability models and parameter estimation methods.
}
\details{
The bootstrap procedure samples from the fitted distribution via inverse
transform sampling. For each bootstrapped sample, the parameters are re-estimated
using the specified \code{method}. Then, the bootstrapped parameters are used to compute
a new set of bootstrapped quantiles. Confidence intervals are obtained from the
empirical nonexceedance probabilities of the bootstrapped quantiles.
}
\note{
The parametric bootstrap is known to give unreasonably wide confidence intervals
for small datasets. If this function detects a confidence interval that is 5+ times
wider than the return levels themselves, it will return an error and recommend
RFPL uncertainty quantification (with \code{uncertainty_rfpl}).
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
uncertainty_bootstrap(data, "WEI", "L-moments")

}
\seealso{
\link{fit_lmom_fast}, \link{fit_maximum_likelihood}, \link{lmom_sample},
\link{quantile_fast}, \link{plot_sffa}, \link{plot_nsffa}
}
