% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-zstatistic.R
\name{z.selection}
\alias{z.selection}
\title{Z-Statistic Method for Distribution Selection}
\usage{
z.selection(data, samples = 10000)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{samples}{Integer (1); the number of bootstrap samples (default is 10000).}
}
\value{
List; results of distribution selection:
\itemize{
\item \code{method}: \code{"Z-selection"}
\item \code{params}: Kappa distribution parameters for the raw AMS data.
\item \code{log_params}: Kappa distribution parameters for the log-transformed AMS data.
\item \code{bootstrap}: Bias and standard deviation of the estimated L-kurtosis.
\item \code{distance}: List of computed Z-statistics for each candidate distribution.
\item \code{recommendation}: Name of the distribution with the smallest Z-statistic.
}
}
\description{
Selects the best-fit distribution by computing a bias-corrected Z-statistic for the sample
L-kurtosis (\eqn{\tau_4}) against the theoretical L-moments for a set of candidate
distributions. The distribution with the smallest absolute Z-score is selected.
}
\details{
The method performs model selection using both raw and log-transformed data. The
distributions which use the raw AMS data are GEV, GLO, PE3, GNO, and WEI. The LP3
distribution uses log-transformed data.

The Z-statistic is determined by fitting a four-parameter Kappa distribution to the
raw and log-transformed data. Then, bootstrapped samples from this Kappa distribution
The L-moments of these bootstrapped samples are used to estimate the Z-statistic
for each distribution.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
z.selection(data)

}
\seealso{
\link{ld.selection}, \link{lk.selection}, \link{pelkap}, \link{qntxxx}
}
