% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-decomposition.R
\name{ams.decomposition}
\alias{ams.decomposition}
\title{Decompose Annual Maximum Streamflow Data}
\usage{
ams.decomposition(data, years, signature)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{signature}{Character (1); Indicates trend(s) to remove.  Must be one of:
\itemize{
\item \code{NULL}: Do not remove any trends.
\item \code{"10"}: Remove linear trend in the mean.
\item \code{"01"}: Remove linear trend in the variance.
\item \code{"11"}: Remove trends in both the mean and variance.
}}
}
\value{
Numeric; a vector with the same length as \code{data} and \code{years} containing
the decomposed AMS data with the specified trend(s) removed.
}
\description{
Removes trends in the means and/or variances of annual maximum streamflow (AMS)
data using Sen’s trend estimator and a moving‐window estimator of the variance.
Four scenarios are supported:
\enumerate{
\item No trend. The data is returned unmodified.
\item Trend in means only.
\item Trend in variance only.
\item Trends in both means and variance.
}
}
\details{
Internally, the function does the following:
\enumerate{
\item If \code{signature == NULL}, returns \code{data} without modification.
\item Constructs a \code{covariate} based on \code{years} using the formula \code{(years - 1900) / 100}.
\item If \code{signature == "10"}, fits Sen’s trend estimator to \code{data} and \code{covariate}
and removes the fitted linear mean trend.
\item If \code{signature == "01"}, computes moving‐window variances using \link{mw.variance},
fits Sen’s trend estimator to those variances, and rescales the series to remove
trends in the variance.
\item If \code{signature == "11"}, applies (3) then (4) sequentially.
\item Ensures all returned values are greater than 1 by shifting the decomposed data.
}
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
ams.decomposition(data, years, "10")

}
\seealso{
\link{mw.variance}, \link{sens.trend}
}
