% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bbmk-test.R
\name{bbmk.plot}
\alias{bbmk.plot}
\title{Plot Block‐Bootstrap Mann–Kendall Test Results}
\usage{
bbmk.plot(results)
}
\arguments{
\item{results}{List; BB‐MK test results generated by \link{bbmk.test}.}
}
\value{
\code{ggplot}; A plot containing:
\itemize{
\item A gray histogram of the distribution of bootstrapped S‐statistics.
\item A red vertical line at the lower and upper confidence bounds.
\item A black vertical line at the observed S‐statistic.
}
}
\description{
Generates a histogram of bootstrapped Mann–Kendall S‐statistics with vertical
lines indicating the observed S‐statistic and confidence bounds.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
results <- bbmk.test(data, samples = 1000)
bbmk.plot(results)

}
