% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-bbmk-test.R
\name{bbmk.test}
\alias{bbmk.test}
\title{Block-Bootstrap Mann-Kendall Test for Trend Detection}
\usage{
bbmk.test(data, alpha = 0.05, samples = 10000, quiet = TRUE)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}

\item{samples}{Integer (1); the number of bootstrap samples (default is 10000).}

\item{quiet}{Logical (1); if FALSE, prints a summary of results (default is TRUE).}
}
\value{
List; the results of the test, including:
\itemize{
\item \code{s.bootstrap}: Vector of bootstrapped test statistics used for plotting.
\item \code{s.statistic}: The Mann-Kendall test statistic computed on the original series.
\item \code{p.value}: Empirical two-sided p-value computed from the bootstrap distribution.
\item \code{bounds}: Confidence interval bounds for the null distribution of the statistic.
\item \code{reject}: Logical. TRUE if the null hypothesis was rejected at significance level alpha.
\item \code{msg}: Character string summarizing the test result (printed if \code{quiet = FALSE}).
}
}
\description{
Block-Bootstrap Mann-Kendall Test for Trend Detection
}
\details{
The block size for the bootstrap is selected as \code{least_lag + 1}, where \code{least_lag}
is estimated using \link{spearman.test}. Each bootstrap sample is generated by
resampling blocks of the original data  (without replacement) and computing the
Mann-Kendall S-statistic. This procedure adjusts for autocorrelation in the data.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
bbmk.test(data, samples = 1000)

}
\seealso{
\link{mk.test}, \link{spearman.test}
}
