% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmk-test.R
\name{bbmk.test}
\alias{bbmk.test}
\title{Block-Bootstrap Mann-Kendall Test for Trend Detection}
\usage{
bbmk.test(data, alpha = 0.05, n_sim = 10000, parallel = FALSE, quiet = TRUE)
}
\arguments{
\item{data}{A numeric vector of annual maximum streamflow data without  NA values.}

\item{alpha}{A numeric value indicating the significance level (default is 0.05).}

\item{n_sim}{An integer number of bootstrap simulations (default is 10000).}

\item{parallel}{Logical. If TRUE, runs the bootstrap in parallel (default is FALSE).}

\item{quiet}{Logical. If FALSE, prints a summary of results (default is TRUE).}
}
\value{
A named list containing:
\describe{
\item{s.bootstrap}{Vector of bootstrapped test statistics used for plotting.}
\item{s.statistic}{The Mann-Kendall test statistic computed on the original series.}
\item{p.value}{Empirical two-sided p-value computed from the bootstrap distribution.}
\item{bounds}{Confidence interval bounds for the null distribution of the statistic.}
\item{reject}{Logical. TRUE if the null hypothesis was rejected at significance level alpha.}
\item{msg}{Character string summarizing the test result (printed if \code{quiet = FALSE}).}
}
}
\description{
Performs a block-bootstrap version of the Mann-Kendall trend test to account
for serial correlation in annual maximum streamflow (AMS) data. The procedure
uses Spearman’s test to estimate the effective block size, then applies a
parallelized bootstrap procedure to obtain the empirical p-value and confidence
bounds for the Mann-Kendall statistic.
}
\details{
The block size for the bootstrap is selected as \code{least_lag + 1}, where
\code{least_lag} is estimated using Spearman’s rank autocorrelation test.
This addresses potential autocorrelation in the AMS data. Each bootstrap sample
is generated by resampling blocks of the original data (without replacement) and
computing the Mann-Kendall statistic.

Using \code{parallel = TRUE} can reduce computation time by approximately 50\%.
However, it is not possible to use \code{set.seed()} if the test is running
in parallel, so your results may not be reproducible.
}
\seealso{
\code{\link{mk.test}}, \code{\link{spearman.test}}
}
