% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN-08NM050.R
\docType{data}
\name{CAN_08NM050}
\alias{CAN_08NM050}
\title{CAN-08NM050}
\format{
A dateframe with 97 rows and 2 columns spanning the period 1921-2017.
}
\source{
Meteorological Service of Canada (MSC) GeoMet Platform
}
\usage{
CAN_08NM050
}
\description{
A dataframe of annual maximum series observations for
station 08NM050, OKANAGAN RIVER AT PENTICTON in British Columbia, Canada.
}
\details{
Variables:
\itemize{
\item \code{max}: Numeric; the annual maximum series observation, in m\eqn{^3}{^3}/s.
\item \code{year}: Integer; the corresponding year.
}
}
\section{Additional Information}{

This is a \emph{regulated} station that is not part of the RHBN. Other notable features include:
\itemize{
\item The Okanagan River upstream of the station has been regulated since 1914 due to the
construction of the first dam, followed by a second dam in 1920, and a regulation system
in the early 1950s, consisting of four dams and 38 km of engineered channel.
\item Rapid human settlement, development, and agricultural activity have occurred in the watershed.
}

This dataset is used as a test case for comparison with the MATLAB implementation of the
FFA framework. It is also useful for demonstrating how the framework detects and handles
serial correlation, trends in the mean, and trends in the variability. As noted above,
this dataset is heavily influenced by reservoir operations and is intended for teaching
purposes—not for design flood estimation.
}

\keyword{datasets}
