% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-pettitt-test.R
\name{eda_pettitt_test}
\alias{eda_pettitt_test}
\title{Pettitt Test for Abrupt Changes in the Mean of a Time Series}
\usage{
eda_pettitt_test(data, years, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
A list containing the test results, including:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{years}: The \code{years} argument.
\item \code{u_t}: Numeric vector of absolute U-statistics for all time indices.
\item \code{k_statistic}: Numeric scalar. The maximum absolute U-statistic.
\item \code{k_critical}: Numeric scalar. The critical K-statistic value for given \code{alpha}.
\item \code{p_value}: Numeric scalar. Approximate p-value for the test.
\item \code{change_index}: Integer scalar. Index of the detected change point (0 if none).
\item \code{change_year}: Integer scalar. Year of the detected change point (0 if none).
\item \code{reject}: Logical scalar. If \code{TRUE}, the null hypothesis was rejected.
\item \code{msg}: Character scalar. A formatted summary message describing the test result.
}
}
\description{
Performs the nonparametric Pettitt test to detect a single abrupt shift in the
mean of a time series. Under the null hypothesis, there is no change point.
}
\details{
The Pettitt test computes the maximum absolute value of the U-statistic
over all possible split points. The p-value is approximated using an
asymptotic formula.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
eda_pettitt_test(data, years)

}
\references{
Pettitt, A.N., 1979. A Non-parametric Approach to the Change-point Problem. J.
Royal Statist. Soc. 28 (2), 126–135. \url{http://www.jstor.org/stable/2346729}
}
\seealso{
\code{\link[=plot_pettitt_test]{plot_pettitt_test()}}
}
