% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-runs-test.R
\name{eda_runs_test}
\alias{eda_runs_test}
\title{Wald–Wolfowitz Runs Test for Randomness}
\usage{
eda_runs_test(results, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{results}{A fitted linear model produced by \link{eda_sens_trend}.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
A list containing the test results, including:
\itemize{
\item \code{residuals}: Numeric vector of residual values from a fitted linear model.
\item \code{n}: The length of the residuals vector after removing the median.
\item \code{n_plus}, \code{n_minus}: The number of residuals above/below the median.
\item \code{runs}: The number of runs in the transformed sequence of residuals.
\item \code{statistic}: The runs test statistic, computed using \code{runs}.
\item \code{p_value}: P-value from the Wald–Wolfowitz runs test applied to residuals.
\item \code{reject}: Logical. If TRUE, the null hypothesis of random residuals is rejected.
\item \code{msg}: Character string summarizing the test result, printed if \code{quiet = FALSE}.
}
}
\description{
Applies the Wald–Wolfowitz runs test to a numeric vector of residuals in
order to assess whether they behave as a random sequence. The test statistic’s
p-value is compared to the significance level \code{alpha}, and a decision is
returned along with a human-readable summary message.
}
\details{
The Wald–Wolfowitz runs test examines the sequence of residuals to test for
randomness around the median. A small p-value suggests nonrandom clustering,
which may indicate that a linear model is inappropriate for the data.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
sens <- eda_sens_trend(data, years)
eda_runs_test(sens)

}
\references{
Wald, A. and Wolfowitz, J. (1940). On a test whether two samples are from the
same population. Annals of Mathematical Statistics, 11(2), 147–162.
}
\seealso{
\link{plot_runs_test}, \link{eda_sens_trend}
}
