% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-sens-trend.R
\name{eda_sens_trend}
\alias{eda_sens_trend}
\title{Sen's Trend Estimator}
\usage{
eda_sens_trend(data, years, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
A list containing the estimated trend:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{years}: The \code{years} argument.
\item \code{slope}: Median slope of all pairwise data-year combinations.
\item \code{intercept}: Median intercept estimate of the fitted line.
\item \code{residuals}: Vector of residuals between observed and fitted values.
\item \code{msg}: Character string summarizing the results.
}
}
\description{
Computes Sen's linear trend estimator for a univariate time series.
}
\details{
Sen's slope estimator is a robust, nonparametric trend estimator based on the
median of all pairwise slopes between data points. The corresponding intercept
is the median of each \eqn{y_i - mx_i} where \eqn{m} is the estimated slope.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
eda_sens_trend(data, years)

}
\references{
Sen, P.K. (1968). Estimates of the regression coefficient based on Kendall's tau.
\emph{Journal of the American Statistical Association}, 63(324), 1379–1389.
}
\seealso{
\code{\link[=eda_runs_test]{eda_runs_test()}}, \code{\link[=plot_sens_trend]{plot_sens_trend()}}
}
