% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-spearman-test.R
\name{eda_spearman_test}
\alias{eda_spearman_test}
\title{Spearman Test for Autocorrelation}
\usage{
eda_spearman_test(data, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{quiet}{Logical scalar. If \code{FALSE}, prints a summary of of the statistical
test to the console. Default is \code{TRUE}.}
}
\value{
A list containing the test results, including:
\itemize{
\item \code{data}: The \code{data} argument.
\item \code{rho}: Numeric vector of serial correlation estimates for lags \eqn{1} to \eqn{n-3}.
\item \code{sig}: Logical vector indicating which lags exhibit significant serial correlation
\item \code{least_lag}: The smallest lag at which the serial correlation is insignificant.
\item \code{reject}: Logical. If \code{TRUE}, then \code{least_lag > 0}.
\item \code{msg}: Character string summarizing the test result, printed if \code{quiet = FALSE}.
}
}
\description{
Performs the Spearman rank serial correlation test on annual maximum series
data to check for serial correlation at various lags. Reports the smallest
lag where the serial correlation is not statistically significant at the given
significance level.
}
\details{
For each lag from \eqn{1} to \eqn{n - 3}, the function computes the Spearman
correlation coefficient between the original series and the lagged series. The
first lag with an insignificant serial correlation coefficient returned as \code{least_lag}.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
eda_spearman_test(data)

}
\seealso{
\code{\link[stats:cor.test]{stats::cor.test()}}, \code{\link[=eda_bbmk_test]{eda_bbmk_test()}}, \code{\link[=plot_spearman_test]{plot_spearman_test()}}
}
