% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-fast.R
\name{fit_lmom_fast}
\alias{fit_lmom_fast}
\title{Helper Function for L-moments Parameter Estimation}
\usage{
fit_lmom_fast(data, model)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}
}
\value{
A numeric vector of parameters:
\itemize{
\item If \code{model} is \code{"GUM"}, \code{"NOR"}, or \code{"LNO"}, returns a vector of length 2.
\item Otherwise, returns a vector of length 3.
}
}
\description{
A helper function used by \link{fit_lmom_methods}.
This function does not validate parameters and is intended for internal use.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmom_fast(data, "PE3")

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\link{lmom_sample}, \link{fit_lmom_kappa}, \link{fit_lmom_methods}
}
