% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lmom-kappa.R
\name{fit_lmom_kappa}
\alias{fit_lmom_kappa}
\title{L-Moments Parameter Estimation for the Kappa Distribution}
\usage{
fit_lmom_kappa(data)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}
}
\value{
A numeric vector of four parameters in the order location, scale,
shape (1), shape (2).
}
\description{
This functions estimates the parameters of the four-parameter Kappa distribution
using the method of L-moments. Since there is no known closed form solution for
the parameters in terms of the L-moments, the parameters are computed numerically
using Newton-Raphson iteration.
}
\details{
First, the sample L-moments of the data are computed using the \link{lmom_sample}
method. Then, the \link[stats]{optim} function is used to determine the
parameters by minimizing the \code{sumquad.tau3tau4} helper function. The implementation of
this routine is based on the deprecated \code{homtest} package.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
fit_lmom_kappa(data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\link{lmom_sample}, \link{fit_lmom_fast}, \link{fit_lmom_methods}
}
