% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-ffa.R
\name{framework_ffa}
\alias{framework_ffa}
\title{Orchestrate Flood Frequency Analysis}
\usage{
framework_ffa(data, years, ns_splits = NULL, ns_structures = NULL)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{ns_splits}{An integer list of years at which to split the data. Use \code{splits = integer(0)}
to avoid splitting the data. Note that \code{integer(0)} is NOT the same as \code{NULL}.}

\item{ns_structures}{An list of structure objects of size \code{length(splits) + 1}. Each
structure object is a list with boolean items \code{location} and \code{scale} indicating a trend
in the mean/variability respectively.}

\item{...}{Additional arguments. See the "Optional Arguments" section for a complete list.}
}
\value{
A list of submodules containing the results of frequency analysis. Each list contains
a \code{name}, which is either \code{"Distribution Selection"}, \code{"Parameter Estimation"},  \code{"Uncertainty  Quantification"}, or \code{"Model Assessment"}. Then, the submodule will contain a sublist of
the results for each subperiod.
}
\description{
Performs frequency analysis of annual maximum series (AMS) data including
distribution selection, parameter estimation, uncertainty quantification,
and model assessment. Supports both stationary frequency analysis (S-FFA)
and nonstationary frequency analysis (NS-FFA).
}
\section{Optional Arguments}{

\itemize{
\item \code{selection}: Distribution selection method (default is \code{"L-distance"}). Must be one of
\code{"L-distance"}, \code{"L-kurtosis"} or \code{"Z-statistic"}. Alternatively, set \code{selection}
to a three-letter distribution code (eg. \code{"GUM"}) to use a specific distribution.
\item \code{s_estimation}: Parameter estimation method for S-FFA (default is \code{"L-moments"}). Must be
\code{"L-moments"}, \code{"MLE"}, or \code{"GMLE"}. Method \code{"GMLE"} requires \code{selection = "GEV"}.
\item \code{ns_estimation}: Parameter estimation method for NS-FFA (default is \code{"MLE"}). Must be
\code{"MLE"} or \code{"GMLE"}. Method \code{"GMLE"} requires \code{selection = "GEV"}.
\item \code{s_uncertainty}: Uncertainty quantification method for S-FFA (default is \code{"Bootstrap"}).
Must be one of \code{"Bootstrap"}, \code{"RFPL"}, or \verb{RFGPL"}. Using method \code{"RFPL"} requires
\code{s_estimation = "MLE"} and method \code{"RFGPL"} requires \code{s_estimation = "GMLE"}.
\item \code{ns_uncertainty}: Uncertainty quantification method for NS-FFA (default is \code{"RFPL"}).
Must be one of \code{"Bootstrap"}, \code{"RFPL"}, or \verb{RFGPL"}. Using method \code{"RFPL"} requires
\code{ns_estimation = "MLE"} and method \code{"RFGPL"} requires \code{ns_estimation = "GMLE"}.
\item \code{z_samples}: Integer number of bootstrap samples for selection method \code{"Z-statistic"}
(default is 10000).
\item \code{gev_prior}: Parameters for the prior distribution of the shape parameter of the GEV
distribution (default is 6, 9). Used with estimation method \code{"GMLE"}.
\item \code{return_periods}: Integer list of return periods (in years) for estimating return levels.
Uses the 2, 5, 10, 20, 50, and 100 year return periods by default.
\item \code{slices}: Integer vector of years at which to estimate the return levels for nonstationary
models. Slices outside of the period will be ignored (default is 1900, 1950, 2000).
\item \code{sb_samples}: Integer number of samples for uncertainty quantification method \code{"Bootstrap"}
(default is 10000).
\item \code{rfpl_tolerance}: Log-likelihood tolerance for uncertainty quantification method \code{"RFPL"}
(default is 0.01).
\item \code{pp_formula}: Plotting position formula for model assessment. Must be one of:
\itemize{
\item "Weibull" (default): \eqn{i / (n + 1)}
\item "Blom": \eqn{(i - 0.375) / (n + 0.25)}
\item "Cunnane": \eqn{(i - 0.4) / (n + 0.2)}
\item "Gringorten": \eqn{(i - 0.44) / (n + 0.12)}
\item "Hazen": \eqn{(i - 0.5) / n}
}
}
}

\seealso{
\code{\link[=select_ldistance]{select_ldistance()}}, \code{\link[=select_lkurtosis]{select_lkurtosis()}}, \code{\link[=select_zstatistic]{select_zstatistic()}},
\code{\link[=fit_lmoments]{fit_lmoments()}}, \code{\link[=fit_mle]{fit_mle()}}, \code{\link[=fit_gmle]{fit_gmle()}}, \code{\link[=uncertainty_bootstrap]{uncertainty_bootstrap()}},
\code{\link[=uncertainty_rfpl]{uncertainty_rfpl()}}, \code{\link[=uncertainty_rfgpl]{uncertainty_rfgpl()}}, \code{\link[=model_assessment]{model_assessment()}}
}
