% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglik-gev.R
\name{general_loglik_gev}
\alias{general_loglik_gev}
\title{Generalized Log-Likelihood Functions for GEV Models}
\usage{
general_loglik_gev(data, params, prior, years = NULL, structure = NULL)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{distribution = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{structure}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain two logical scalars:
\itemize{
\item \code{location}: If \code{TRUE}, the location parameter has a linear temporal trend.
\item \code{scale}: If \code{TRUE}, the scale parameter has a linear temporal trend.
}}
}
\value{
Numeric scalar. The generalized log-likelihood value.
}
\description{
Computes the generalized log-likelihood for stationary and nonstationary
variants of the Generalized Extreme Value (GEV) distribution with a geophysical
(Beta) prior distribution for the shape parameter (Martins and Stedinger, 2000).
}
\details{
The generalized log-likelihood is defined as sum of the log-likelihood of the
specified model and the log-density of the Beta prior with parameters \eqn{(p, q)}.
The contribution of the prior is: \deqn{\log \pi(\kappa) = (p-1) \log(0.5-\kappa) 
+ (q-1) \log(0.5+\kappa) - \log (\beta(p, q))}
}
\examples{
# Initialize data, params, prior, years, and structure
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0)
prior <- c(5, 10)
years <- seq(from = 1901, to = 2000)
structure <- list(location = TRUE, scale = FALSE)

# Compute the generalized log-likelihood
general_loglik_gev(data, params, prior, years, structure)

}
\references{
El Adlouni, S., Ouarda, T.B.M.J., Zhang, X., Roy, R., Bob´ee, B., 2007. Generalized
maximum likelihood estimators for the nonstationary generalized extreme value
model. Water Resour. Res. 43 (3), 1–13. \doi{10.1029/2005WR004545}

Martins, E. S., and Stedinger, J. R. (2000). Generalized maximum-likelihood generalized
extreme-value quantile estimators for hydrologic data. Water Resources Research, 36(3),
737–744. \doi{10.1029/1999WR900330}
}
\seealso{
\code{\link[=general_loglik_fast]{general_loglik_fast()}}
}
