% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglik-fast.R, R/gll-xxx.R
\name{gllxxx}
\alias{gllxxx}
\title{Generalized Log-Likelihood Helper Function}
\usage{
gllxxx(model, trend, data, params, prior, years = NULL)

gllxxx(model, trend, data, params, prior, years = NULL)
}
\arguments{
\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{prior}{Numeric (2); a vector of parameters \eqn{(p, q)} of the Beta prior
on \eqn{\kappa}.}

\item{years}{Numeric; a vectorof years with the same length as \code{data}.
Required if \code{trend$location} or ic    \code{trend$scale} is TRUE.}
}
\value{
Numeric (1); the generalized log-likelihood value.

Numeric (1); the generalized log-likelihood value.
}
\description{
A helper function used by \link{gllgev}, \link{gllgev100}, and \link{gllgev110}.

A helper function used by \link{gllgev}, \link{gllgev100}, and \link{gllgev110}.
}
\note{
This function does not perform parameter validation, which improves performance
but may cause unpredictable behaviour. Use at your own risk.

This function does not perform parameter validation, which improves performance
but may cause unpredictable behaviour. Use at your own risk.
}
\examples{
# Initialize data, params, params, prior, and years
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0, 0)
prior <- c(5, 10)
years <- seq(from = 1901, to = 2000)

# Compute the generalized log-likelihood
gllxxx("GEV", "11", data, params, prior, years)

# Initialize data, params, params, prior, and years
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 1, 0, 0)
prior <- c(5, 10)
years <- seq(from = 1901, to = 2000)

# Compute the generalized log-likelihood
gllxxx("GEV", "11", data, params, prior, years)

}
\seealso{
\link{gll-functions}

\link{gll-functions}
}
