% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpss-test.R
\name{kpss.test}
\alias{kpss.test}
\title{Kwiatkowski–Phillips–Schmidt–Shin (KPSS) Unit Root Test}
\usage{
kpss.test(data, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}

\item{quiet}{Logical (1); if FALSE, prints a summary of results (default is TRUE).}
}
\value{
List; the test results, consisting of:
\itemize{
\item \code{statistic}: The KPSS test statistic.
\item \code{p.value}: The reported p-value from the test. See notes regarding discrete thresholds.
\item \code{reject}: Logical. TRUE if the null hypothesis is rejected at \code{alpha}.
\item \code{msg}: Character string summarizing the test outcome, printed if \code{quiet = FALSE}.
}
}
\description{
Performs the KPSS unit root test on annual maximum streamflow (AMS) data.
The null hypothesis is that  the time series is trend-stationary with a linear
trend and constant drift.  The alternative hypothesis is that the time series
has a unit root and is non-stationary.
}
\details{
The implementation of the KPSS test is based on the \pkg{aTSA} package, which interpolates
a significance table from Hobjin et al. (2004). Therefore, a result of \eqn{p = 0.01}
implies that \eqn{p \leq 0.01} and a result of \eqn{p = 0.10} implies that \eqn{p \geq 0.10}.
This implementation uses the Type III KPSS test, which accounts for a linear trend in the data.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
kpss.test(data)

}
\references{
Hobijn, B., Franses, P.H. and Ooms, M. (2004), Generalizations of the KPSS-test for
stationarity. Statistica Neerlandica, 58: 483-502.
}
\seealso{
\link{pp.test}
}
