% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk-selection.R
\name{lk.selection}
\alias{lk.selection}
\title{L-Kurtosis Method for Distribution Selection}
\usage{
lk.selection(data)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}
}
\value{
List; results of distribution selection:
\itemize{
\item \code{method}: \code{"L-kurtosis"}
\item \code{metrics}: A list of L-kurtosis metrics for each distribution.
\item \code{recommendation}: Name of the distribution with the smallest L-kurtosis metric
}
}
\description{
Selects a probability distribution by minimizing the absolute distance
between the theoretical L-kurtosis (\eqn{\tau_4}) and the sample L-kurtosis
(\eqn{t_4}). For 3-parameter distributions, we use the shape parameter that
best replicates the sample L-skewness (\eqn{t_3}) of the data.
}
\details{
This method computes the distance between the sample and theoretical L-kurtosis values at
a fixed L-skewness. For three parameter distributions, the shape parameter that best
replicates the sample L-skewness is derived using \link[stats]{optim}.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
lk.selection(data)

}
\seealso{
\link{lmom.sample}, \link{ld.selection}, \link{z.selection}, \link[stats]{optim}
}
