% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llv-xxx.R
\name{llvxxx}
\alias{llvxxx}
\title{Log-Likelihood Helper Function}
\usage{
llvxxx(name, signature, data, params, covariate = NULL)
}
\arguments{
\item{name}{Character (1); the name of the probability distribution.}

\item{signature}{Character (1); the non-stationary signature (\code{NULL}, \code{"10"}, or \code{"11"}).}

\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{covariate}{Numeric; a vector with the same length as \code{data}.
Required if \code{signature} is \code{"10"} or \code{"11"}.}
}
\value{
Numeric (1); the log-likelihood value.
}
\description{
A helper function used by \link{llv-functions}.
}
\note{
This function does not perform parameter validation, which improves performance
at the cost of potentially unpredictable behaviour. Use at your own risk.
}
\examples{
# Initialize data and params
data <- rnorm(n = 100, mean = 100, sd = 10)
params <- c(0, 1, 0)

# Compute the log-likelihood
llvxxx("GEV", NULL, data, params)

}
\seealso{
\link{llv-functions}
}
