% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lmom-diagram.R
\name{lmom.plot}
\alias{lmom.plot}
\title{Plot L-Moment Ratio Diagram}
\usage{
lmom.plot(data, results)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{results}{List; output of \link{ld.selection}, \link{lk.selection},
or \link{z.selection}.}
}
\value{
\code{ggplot}; plot object containing the L-moment ratio diagram, with:
\itemize{
\item L-moment ratio curves for each 3-parameter distribution.
\item Points for the L-moment ratios of each 2-parameter distribution.
\item Sample and log-sample L-moment ratio \eqn{(t_3, t_4)} points.
}
}
\description{
Generates a plot of L-moment ratios with the L-skewness on the x-axis and L-kurtosis
on the y-axis. Plots the sample and log-sample L-moment ratios alongside the theoretical
L-moment ratios for a set of candidate distributions. If the selection method is
\link{ld.selection} or \link{lk.selection}, the plot will include a small inset around
the L-moment ratios of the recommended distribution.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
results <- ld.selection(data)
lmom.plot(data, results)

}
\seealso{
\link{ld.selection}, \link{lk.selection}, \link{z.selection}
}
