% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmom-fast.R
\name{lmom_fast}
\alias{lmom_fast}
\title{Helper Function for L-moments Ratios}
\usage{
lmom_fast(distribution, params)
}
\arguments{
\item{distribution}{A three-character code indicating the distribution family.
Must be \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"},
or \code{"WEI"}.}

\item{params}{Numeric vector of distribution parameters, in the order (location,
scale, shape). The length must be between 2 and 5, depending on the specified
\code{distribution} and \code{structure}.}
}
\value{
A numeric vector of with four elements:
\itemize{
\item \eqn{\lambda_1}: L-mean
\item \eqn{\lambda_2}: L-variance
\item \eqn{\tau_3}: L-skewness
\item \eqn{\tau_4}: L-kurtosis
}
}
\description{
A helper function used by \code{\link[=lmom_theoretical]{lmom_theoretical()}}.
This function does not validate parameters and is designed for use in other methods.
}
\note{
This function returns identical L-moment ratios for \code{"NOR"}/\code{"LNO"} and
\code{"PE3"}/\code{"LP3"} since L-moments for the \code{"LNO"} and \code{"LP3"} distributions are
compared with the sample L-moments of log-transformed data internally.
}
\examples{
lmom_fast("GLO", c(0, 1, 0))

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\code{\link[=lmom_theoretical]{lmom_theoretical()}}
}
