% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-estimation.R
\name{mle.estimation}
\alias{mle.estimation}
\title{Maximum Likelihood Parameter Estimation}
\usage{
mle.estimation(data, years, model, prior = NULL)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{model}{Character (1); string specifying the probability model. The first three
letters denote the family: \code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3},
\code{LP3}, or \code{WEI}. A trailing signature of \code{10} or \code{100} indicates a linear trend
in location; \code{11} or \code{110} indicates linear trends in both location and scale.}

\item{prior}{Numeric (2); optional vector of parameters \eqn{(p, q)} that specifies
the parameters of a Beta prior on \eqn{\kappa}. Only works with models \code{GEV},
\code{GEV100}, and \code{GEV110}.}
}
\value{
List; results of parameter estimation:
\itemize{
\item \code{params}: Numeric vector of estimated parameters.
\item \code{mll}: Maximum log‐likelihood value.
}
}
\description{
Estimates parameters of a probability distribution (\code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO},
\code{GNO}, \code{PE3}, \code{LP3}, or \code{WEI}) or non-stationary variant by maximizing the log‐likelihood.
Initial values are obtained through L‐moment parameter estimation, and optimization is
performed via \link[stats]{nlminb} with repeated perturbations if needed.
}
\details{
\enumerate{
\item Calls \link{pelxxx} on \code{data} to obtain initial parameter estimates.
\item Initializes trend parameters to zero if there is a trailing signature.
\item For \code{WEI} models, sets the location parameter to zero to ensure support.
\item Defines an objective function as the negative of the \link{llvxxx} function.
\item Runs \link[stats]{nlminb} with box constraints. Attempts optimization
up to 100 times.
}
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
mle.estimation(data, years, "GNO100")

}
\seealso{
\link{llvxxx}, \link{gllxxx}, \link{pelxxx}, \link[stats]{nlminb}
}
