% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{param-data}
\alias{param-data}
\title{Parameter 'data'}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}
}
\description{
Parameter 'data'
}
\keyword{internal}
