% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-data}
\alias{param-data}
\title{Parameter 'data'}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}
}
\description{
Parameter 'data'
}
