% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-ns-splits}
\alias{param-ns-splits}
\title{Parameter 'ns_splits'}
\arguments{
\item{ns_splits}{An integer vector of years used to split the data into homogeneous
subperiods. For S-FFA, set to \code{NULL} (default). For NS-FFA, specify an integer vector
of years (e.g., \code{1950L}) with physical justification for change points, or \code{NULL}
if no such years exist.}
}
\description{
Parameter 'ns_splits'
}
\keyword{internal}
