% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-prior}
\alias{param-prior}
\title{Parameter 'prior'}
\arguments{
\item{prior}{Numeric vector of length 2. Specifies the parameters of the
Beta prior for the shape parameter \eqn{\kappa}.}
}
\description{
Parameter 'prior'
}
\keyword{internal}
