% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffaframework-parameters.R
\name{param-prior}
\alias{param-prior}
\title{Parameter 'prior'}
\arguments{
\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{distribution = "GEV"}.}
}
\description{
Parameter 'prior'
}
\keyword{internal}
