% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-samples}
\alias{param-samples}
\title{Parameter 'samples'}
\arguments{
\item{samples}{Integer scalar. The number of bootstrap samples. Default is 10000.}
}
\description{
Parameter 'samples'
}
