% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-ffa-framework.R
\name{param-slice}
\alias{param-slice}
\title{Parameter 'slice'}
\arguments{
\item{slice}{Numeric scalar specifying the year at which to evaluate the
quantiles or confidence intervals of a non-stationary probability distribution.
The year does not have to be an element of the \code{years} argument. Note that if
\code{trend$location} and \code{trend$scale} are both \code{FALSE}, this argument will have
no effect the output of the function.}
}
\description{
Parameter 'slice'
}
\keyword{internal}
