% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ams-data.R
\name{plot_ams_data}
\alias{plot_ams_data}
\title{Plot Annual Maximum Series Data}
\usage{
plot_ams_data(data, years, show_line = TRUE, ...)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{show_line}{If \code{TRUE} (default), draw a line through the data.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; A plot containing:
\itemize{
\item Black dots at each (data, year) pair.
\item An optional line thorugh the data points (if \code{show_line == TRUE})
}
}
\description{
Generates a plot of annual maximum series data with an optional line connecting
the data points.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
plot_ams_data(data, years)

}
