% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sens-trend.R
\name{plot_sens_trend}
\alias{plot_sens_trend}
\title{Plot Sen’s Trend Estimator}
\usage{
plot_sens_trend(results, type, show_trend = TRUE, ...)
}
\arguments{
\item{results}{A list of Sen's trend estimates generated by \link{eda_sens_trend}.}

\item{type}{A character scalar representing whether Sen's trend estimator was
applied directly to the AMS data (\code{"mean"}) or to the moving-window variance
of the data (\code{"variance"}). Must be either \code{"variance"} or \code{"mean"}.}

\item{show_trend}{If \code{TRUE} (default), draw a fitted line through the AMS data.}

\item{...}{Optional named arguments: 'title', 'xlabel', and 'ylabel'.}
}
\value{
\code{ggplot}; a plot containing:
\itemize{
\item Black points for each year’s AMS (or variance) value.
\item Optional black line connecting the AMS data if \code{show_trend = TRUE}.
\item Blue line representing Sen’s trend estimator.
\item A text annotation displaying the fitted equation \eqn{y = mx + b}.
}
}
\description{
Produces a scatter plot of the annual maximum streamflow (AMS) data or its variance
against time, overlaid with Sen’s trend estimator and an annotation of the fitted
equation.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
results <- eda_sens_trend(data, years)
plot_sens_trend(results, "mean")

}
\seealso{
\link{eda_sens_trend}
}
