% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-test.R
\name{pp.test}
\alias{pp.test}
\title{Phillips–Perron Unit Root Test}
\usage{
pp.test(data, alpha = 0.05, quiet = TRUE)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow data with no missing values.}

\item{alpha}{Numeric significance level. Must be greater than 0.01 (default is 0.05).}

\item{quiet}{Logical. If FALSE, prints a summary message to the console (default is TRUE).}
}
\value{
A named list containing:
\describe{
\item{statistic}{The Z-statistic used to perform the test.}
\item{p.value}{Reported p-value from the test. See notes on interpolation thresholds.}
\item{reject}{Logical. TRUE if the null hypothesis of a unit root is rejected at \code{alpha}.}
\item{msg}{Character string summarizing the test result (printed if \code{quiet = FALSE}).}
}
}
\description{
Applies the Phillips–Perron (PP) test to assess the presence of a unit root in annual
maximum streamflow (AMS) data. The null hypothesis is that the series contains a unit root
(i.e., is non-stationary). This implementation of the PP test assumes the time series
has a stationary drift term and linear trend term.
}
\details{
The implementation of this test is based on the \pkg{aTSA} package, which interpolates p-values
\code{alpha > 0.01}. Note that a reported p-value of 0.01 indicates \eqn{p \leq 0.01}.

The null hypothesis is that the time series contains a unit root (non-stationary). Rejection
of the null suggests that the series is trend-stationary.
}
\references{
Fuller, W. A. (1996). Introduction to statistical time series, second ed., Wiley, New York.
}
\seealso{
\code{\link[aTSA]{pp.test}}, \code{\link{kpss.test}}
}
