% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile-fast.R
\name{qntxxx}
\alias{qntxxx}
\title{Helper Function for Quantile Functions}
\usage{
qntxxx(model, trend, p, params, years = 0)
}
\arguments{
\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{p}{Numeric; a vector of probabilities. Must be between 0 and 1.}

\item{params}{Numeric; a vector of parameters. Must have the correct length for the model.}

\item{years}{Numeric; a vector of years at which to compute quantiles.
Required if \code{trend$location} or \code{trend$scale} is TRUE.}
}
\value{
If \code{p} or \code{years} is a scalar, returns a numeric vector. Otherwise, returns a matrix.
}
\description{
A helper function used by \link{qnt-functions}.
}
\examples{
# Initialize p and params
p <- runif(n = 10)
params <- c(0, 1, 0)

# Compute the log-likelihood
qntxxx("GEV", NULL, p, params)

}
\seealso{
\link{qnt-functions}
}
