% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfpl-uncertainty.R
\name{rfpl.uncertainty}
\alias{rfpl.uncertainty}
\title{Regula-Falsi Confidence Intervals for Flood Quantile Estimates}
\usage{
rfpl.uncertainty(
  data,
  years,
  model,
  slices = "last",
  alpha = 0.05,
  eps = 0.01,
  prior = NULL
)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{model}{Character (1); string specifying the probability model. The first three
letters denote the family: \code{GUM}, \code{NOR}, \code{LNO}, \code{GEV}, \code{GLO}, \code{GNO}, \code{PE3},
\code{LP3}, or \code{WEI}. A trailing signature of \code{10} or \code{100} indicates a linear trend
in location; \code{11} or \code{110} indicates linear trends in both location and scale.}

\item{slices}{Character (1) or Numeric; specifies the years at which
to compute the estimates and confidence intervals (default is "last").
\itemize{
\item \code{"all"}: returns estimates for all values in \code{years}.
\item \code{"first"}: returns estimates for first year in the dataset.
\item \code{"last"}: returns estimates for last year in the dataset.
\item Passing a numeric vector to \code{slices} allows for custom values.
}

If the chosen model is stationary, the results will be the same for all slices.}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}

\item{eps}{Numeric (1); tolerance for the Regula-Falsi convergence (default is 0.01).}

\item{prior}{Numeric (2); optional vector of parameters \eqn{(p, q)} that specifies
the parameters of a Beta prior on \eqn{\kappa}. Only works with models \code{GEV},
\code{GEV100}, and \code{GEV110}.}
}
\value{
List; quantiles and confidence intervals. Each year maps to a sub-list:
\itemize{
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{t}: Vector of return periods; \code{c(2, 5, 10, 20, 50, 100)}.
}
}
\description{
Calculates estimates and confidence intervals for return levels at standard
return periods (2, 5, 10, 20, 50, and 100 years) using the profile likelihood
and Regula-Falsi root‐finding method.
}
\details{
\enumerate{
\item Fits the model using \link{mle.estimation} to obtain parameter estimates and log‐likelihood.
\item Defines an objective function \eqn{f(y_p, p)} based on the chi-squared distribution.
\item Iteratively brackets the root by scaling initial guesses by 0.05 until f changes sign.
\item Uses the Regula Falsi method to solve \eqn{f(y_p, p) = 0} for each return-period probability.
\item Returns lower and upper confidence bounds at level alpha and the quantile estimates.
}
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
rfpl.uncertainty(data, years, "GLO110")

}
\seealso{
\link{qntxxx}, \link{sb.uncertainty}, \link[stats]{nlminb}
}
