% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-bootstrap.R
\name{sb.uncertainty}
\alias{sb.uncertainty}
\title{Sample Bootstrap Confidence Intervals for Flood Quantile Estimates}
\usage{
sb.uncertainty(
  data,
  years,
  model,
  trend,
  method,
  slices = "last",
  samples = 10000,
  alpha = 0.05,
  prior = NULL
)
}
\arguments{
\item{data}{Numeric; a vector of annual maximum streamflow data.}

\item{years}{Numeric; a vector of years with the same length as \code{data}.}

\item{model}{Character (1); three character distribution code. Must be one of:
\code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{method}{Character (1); string specifying the estimation method.
Must be \code{"L-moments"}, \code{"MLE"}, or \code{"GMLE"}.}

\item{slices}{Character (1) or Numeric; specifies the years at which
to compute the estimates and confidence intervals (default is "last").
\itemize{
\item \code{"all"}: returns estimates for all values in \code{years}.
\item \code{"first"}: returns estimates for first year in the dataset.
\item \code{"last"}: returns estimates for last year in the dataset.
\item Passing a numeric vector to \code{slices} allows for custom values.
}

If the \code{trend$location} and \code{trend$scale} are FALSE, the results will
be the same for all slices.}

\item{samples}{Integer (1); the number of bootstrap samples (default is 10000).}

\item{alpha}{Numeric (1); the significance level (default is 0.05).}

\item{prior}{Numeric (2); optional vector of parameters \eqn{(p, q)} that specifies
the parameters of a Beta prior on \eqn{\kappa}. Only works with distribution \code{"GEV"}.}
}
\value{
List; quantiles and confidence intervals. Each year maps to a sub-list:
\itemize{
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{t}: Vector of return periods; \code{c(2, 5, 10, 20, 50, 100)}.
}
}
\description{
Computes estimates and confidence intervals for return levels at standard return periods
(2, 5, 10, 20, 50, and 100 years) using the sample bootstrap method. This function
supports a variety of probability models and parameter estimation methods.
}
\details{
The bootstrap procedure samples from the fitted distribution via inverse
transform sampling. For each bootstrapped sample, the parameters are re-estimated
using the specified \code{method}. Then, the bootstrapped parameters are used to compute
a new set of bootstrapped quantiles. Confidence intervals are obtained from the
empirical non-exceedance probabilities of the bootstrapped quantiles.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
sb.uncertainty(data, years, "WEI", "L-moments")

}
\seealso{
\link{pelxxx}, \link{mle.estimation}, \link{lmom.sample}, \link[stats]{quantile}
}
