% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-zstatistic.R
\name{select_zstatistic}
\alias{select_zstatistic}
\title{Z-Statistic Method for Distribution Selection}
\usage{
select_zstatistic(data, samples = 10000L)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{samples}{Integer scalar. The number of bootstrap samples. Default is 10000.}
}
\value{
A list with the results of distribution selection:
\itemize{
\item \code{method}: \code{"Z-selection"}.
\item \code{data}: The \code{data} argument.
\item \code{metrics}: List of computed Z-statistics for each candidate distribution.
\item \code{recommendation}: Name of the distribution with the smallest Z-statistic.
\item \code{reg_params}: Kappa distribution parameters for the data.
\item \code{reg_bias_t4}: Bias of the L-kurtosis from the bootstrap.
\item \code{reg_std_t4}: Standard deviation of the L-kurtosis from the bootstrap.
\item \code{log_params}: Kappa distribution parameters for the log-transformed data.
\item \code{log_bias_t4}: Bias of the L-kurtosis from the bootstrap (using \code{log_params}).
\item \code{log_std_t4}: Standard deviation of the L-kurtosis from the bootstrap (using \code{log_params}).
}
}
\description{
Selects the best-fit distribution by comparing a bias-corrected Z-statistic for
the sample L-kurtosis (\eqn{\tau_4}) against the theoretical L-moments for a set
of candidate distributions. The distribution with the smallest absolute Z-statistic
is selected.
}
\details{
The method performs distribution selection using both raw and log-transformed data. The
distributions which use the raw data are GEV, GLO, PE3, GNO, and WEI. The LP3 distribution
uses the log-transformed data.

The Z-statistic is determined by fitting a four-parameter Kappa distribution to the
raw and log-transformed data. Then, bootstrapped samples from this Kappa distribution
The L-moments of these bootstrapped samples are used to estimate the Z-statistic
for each distribution.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
select_zstatistic(data)

}
\references{
Hosking, J.R.M. & Wallis, J.R., 1997. Regional frequency analysis: an approach based
on L-Moments. Cambridge University Press, New York, USA.
}
\seealso{
\code{\link[=select_ldistance]{select_ldistance()}}, \code{\link[=select_lkurtosis]{select_lkurtosis()}}, \code{\link[=fit_lmom_kappa]{fit_lmom_kappa()}},
\code{\link[=quantile_fast]{quantile_fast()}}, \code{\link[=plot_lmom_diagram]{plot_lmom_diagram()}}
}
