% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-zstatistic.R
\name{select_zstatistic}
\alias{select_zstatistic}
\title{Z-Statistic Method for Distribution Selection}
\usage{
select_zstatistic(data, samples = 10000L)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{samples}{Integer scalar. The number of bootstrap samples. Default is 10000.}
}
\value{
A list with the results of distribution selection:
\itemize{
\item \code{method}: \code{"Z-selection"}.
\item \code{data}: The \code{data} argument.
\item \code{metrics}: List of computed Z-statistics for each candidate distribution.
\item \code{recommendation}: Name of the distribution with the smallest Z-statistic.
\item \code{reg_params}: Kappa distribution parameters for the AMS data.
\item \code{reg_bias_t4}: Bias of the L-kurtosis from the AMS bootstrap.
\item \code{reg_std_t4}: Standard deviation of the L-kurtosis from the AMS bootstrap.
\item \code{log_params}: Kappa distribution parameters for the log-AMS data.
\item \code{log_bias_t4}: Bias of the L-kurtosis from the log-AMS bootstrap.
\item \code{log_std_t4}: Standard deviation of the L-kurtosis from the log-AMS bootstrap.
}
}
\description{
Selects the best-fit distribution by computing a bias-corrected Z-statistic for
the sample L-kurtosis (\eqn{\tau_4}) against the theoretical L-moments for a set
of candidate distributions. The distribution with the smallest absolute Z-statistic
is selected.
}
\details{
The method performs model selection using both raw and log-transformed data. The
distributions which use the raw AMS data are GEV, GLO, PE3, GNO, and WEI. The LP3
distribution uses log-transformed data.

The Z-statistic is determined by fitting a four-parameter Kappa distribution to the
raw and log-transformed data. Then, bootstrapped samples from this Kappa distribution
The L-moments of these bootstrapped samples are used to estimate the Z-statistic
for each distribution.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
select_zstatistic(data)

}
\seealso{
\link{select_ldistance}, \link{select_lkurtosis}, \link{fit_lmom_kappa},
\link{quantile_fast}, \link{plot_lmom_diagram}
}
