% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spearman-test.R
\name{spearman.plot}
\alias{spearman.plot}
\title{Plot Spearman’s Rho Autocorrelation}
\usage{
spearman.plot(results)
}
\arguments{
\item{results}{List; autocorrelation coefficients generated by \link{spearman.test}.}
}
\value{
\code{ggplot}; a plot showing:
\itemize{
\item Vertical segments from \eqn{y=0} up to each \eqn{\rho} value at its lag.
\item Filled circles at each lag, filled black if serial correlation is detected.
}
}
\description{
Visualizes Spearman’s rho autocorrelation coefficients with
shaded points indicating statistical significance.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
results <- spearman.test(data)
spearman.plot(results)

}
