% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-uncertainty.R
\name{uncertainty.plot}
\alias{uncertainty.plot}
\title{Plot Return Levels and Confidence Intervals}
\usage{
uncertainty.plot(results, trend, slice = NULL)
}
\arguments{
\item{results}{List; generated by \link{sb.uncertainty} or \link{rfpl.uncertainty}.}

\item{trend}{List; information about non-stationary trend(s) to use:
\itemize{
\item \code{location} Logical (1); if TRUE, there is a trend in the location parameter.
\item \code{scale} Logical (1); if TRUE, there is a trend in the scale parameter.
}}

\item{slice}{Character (1); the anchor year for non-stationary return level estimates.
Required for models ending in \code{10}, \code{100}, \code{11}, or \code{110}.}
}
\value{
\code{ggplot}; a plot showing:
\itemize{
\item A dashed black line for both the lower and upper confidence bounds.
\item A semi-transparent gray ribbon between \code{results$ci_lower} and \code{results$ci_upper}.
\item A solid blue line for the point estimates.
}
}
\description{
Constructs a time–series plot of estimated return levels along with their
associated confidence intervals. The confidence bounds are shown as a shaded ribbon,
and the point estimates are overlaid as a solid line. The return periods are
displayed on a logarithmic scale.

If the specified model is non-stationary, the \code{slice} argument is required
and the return levels are reported as "Effective Return Periods".
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
results <- sb.uncertainty(data, years, "WEI", "L-moments")
uncertainty.plot("WEI", results)

}
