% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-bootstrap.R
\name{uncertainty_bootstrap}
\alias{uncertainty_bootstrap}
\title{Parametric Bootstrap Confidence Intervals for Flood Quantile Estimates}
\usage{
uncertainty_bootstrap(
  data,
  distribution,
  method,
  prior = NULL,
  years = NULL,
  structure = NULL,
  slices = 1900,
  alpha = 0.05,
  samples = 10000L,
  periods = c(2, 5, 10, 20, 50, 100)
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series values.
Must be strictly positive, finite, and not missing.}

\item{distribution}{A three-character code indicating the distribution family.
Must be \code{"GUM"}, \code{"NOR"}, \code{"LNO"}, \code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"},
or \code{"WEI"}.}

\item{method}{Character scalar specifying the estimation method.
Must be \code{"L-moments"}, \code{"MLE"}, or \code{"GMLE"}.}

\item{prior}{Numeric vector of length 2. Specifies the parameters of the
Beta prior for the shape parameter \eqn{\kappa}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{samples}{Integer scalar. The number of bootstrap samples. Default is 10000.}

\item{periods}{Numeric vector used to set the return periods for FFA.
All entries must be greater than or equal to 1.}
}
\value{
A list containing the following three items:
\itemize{
\item \code{method}: "Bootstrap"
\item \code{structure}: The value of the \code{structure} argument.
\item \code{slices}: A list of lists containing the results for each slice.
}

Each element of \code{slices} is a list with the following five items:
\itemize{
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{periods}: Vector of return periods; defaults to \code{c(2, 5, 10, 20, 50, 100)}.
\item \code{year}: The year at which the estimates were computed (nonstationary models only).
}
}
\description{
Computes return level estimates and confidence intervals at the specified return
periods (defaults to 2, 5, 10, 20, 50, and 100 years) using the parametric bootstrap
method. This function supports a variety of probability models and parameter estimation
methods.
}
\details{
The bootstrap procedure samples from the fitted distribution via inverse
transform sampling. For each bootstrapped sample, the parameters are re-estimated
using the \code{method} argument. Then, the bootstrapped parameters are used to compute
a new set of bootstrapped quantiles. Confidence intervals are obtained from the
empirical nonexceedance probabilities of the bootstrapped quantiles.
}
\note{
The parametric bootstrap is known to give unreasonably wide confidence intervals
for small datasets. If this function detects a confidence interval that is 5+ times
wider than the return levels themselves, it will return an error and recommend
RFPL uncertainty quantification (with \code{uncertainty_rfpl}).
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
years <- seq(from = 1901, to = 2000)
uncertainty_bootstrap(data, "WEI", "L-moments")

}
\references{
Vidrio-Sahagún, C.T., He, J. Enhanced profile likelihood method for the nonstationary
hydrological frequency analysis, Advances in Water Resources 161, 10451 (2022).
\doi{10.1016/j.advwatres.2022.104151}
}
\seealso{
\code{\link[=fit_lmom_fast]{fit_lmom_fast()}}, \code{\link[=fit_maximum_likelihood]{fit_maximum_likelihood()}}, \code{\link[=lmom_sample]{lmom_sample()}},
\code{\link[=quantile_fast]{quantile_fast()}}, \code{\link[=plot_sffa]{plot_sffa()}}, \code{\link[=plot_nsffa]{plot_nsffa()}}
}
