% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-rfpl.R
\name{uncertainty_rfpl}
\alias{uncertainty_rfpl}
\title{Regula-Falsi Confidence Intervals for Flood Quantile Estimates}
\usage{
uncertainty_rfpl(
  data,
  model,
  prior = NULL,
  years = NULL,
  trend = NULL,
  slices = 1900,
  alpha = 0.05,
  eps = 0.01,
  periods = c(2, 5, 10, 20, 50, 100)
)
}
\arguments{
\item{data}{Numeric vector of annual maximum series (AMS) values.
Must be strictly positive, finite, and not missing.}

\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{model = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as nonstationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}

\item{slices}{Numeric vector specifying the years at which to evaluate the
return levels confidence intervals of a nonstationary probability distribution.
The slices do not have to be elements of the \code{years} argument.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{eps}{Numeric scalar. The log-likelihood tolerance for the Regula-Falsi
convergence (default is 0.01).}

\item{periods}{Numeric vector used to set the return periods for FFA.
All entries must be greater than or equal to 1.}
}
\value{
A list of lists containing the return levels and confidence
intervals for each slice. Each sub-list contains:
\itemize{
\item \code{method}: Either "RFPL" or "RFGPL"
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{t}: Vector of return periods \code{c(2, 5, 10, 20, 50, 100)} in years.
\item \code{slice}: The value of \code{slice} argument.
\item \code{trend}: The value of \code{trend} argument.
}
}
\description{
Calculates estimates and confidence intervals for return levels at standard
return periods (2, 5, 10, 20, 50, and 100 years) using the profile likelihood
and Regula-Falsi root‐finding method.
}
\details{
\enumerate{
\item Fits the model using \link{fit_maximum_likelihood} to obtain parameter
estimates and log‐likelihood.
\item Defines an objective function \eqn{f(y_p, p)} based on the chi-squared
distribution.
\item Iteratively brackets the root by scaling initial guesses by 0.05 until
\eqn{f(y_p, p)} changes sign.
\item Uses the Regula Falsi method to solve \eqn{f(y_p, p) = 0} for each
return-period probability.
\item Returns lower and upper confidence bounds at significance level \code{alpha}
given the return level estimates.
}
}
\note{
Although the more modern \link[stats]{optim} function is preferred over
\link[stats]{nlminb}, we use \link[stats]{nlminb} because it supports infinite
values of the likelihood function.

RFPL uncertainty quantification can be numerically unstable for some datasets.
If this function encounters an issue, it will return an error and recommend
using the parametric bootstrap (\code{uncertainty_bootstrap}) instead.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
uncertainty_rfpl(data, "GLO")

}
\seealso{
\link{quantile_fast}, \link{uncertainty_bootstrap},
\link{plot_sffa}, \link{plot_nsffa}
}
