% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-rfpl.R
\name{uncertainty_rfpl}
\alias{uncertainty_rfpl}
\title{Regula-Falsi Confidence Intervals for Flood Quantile Estimates}
\usage{
uncertainty_rfpl(
  data,
  model,
  prior = NULL,
  years = NULL,
  trend = NULL,
  slice = NULL,
  alpha = 0.05,
  eps = 0.01
)
}
\arguments{
\item{data}{Numeric vector of annual maximum streamflow (AMS) values.
Must be strictly positive, finite, and non-missing.}

\item{model}{Character scalar. A three-character code indicating
a distribution family. Must be one of:  \code{"GUM"}, \code{"NOR"}, \code{"LNO"},
\code{"GEV"}, \code{"GLO"}, \code{"GNO"}, \code{"PE3"}, \code{"LP3"}, or \code{"WEI"}.}

\item{prior}{Numeric vector of length 2. Specifies the Beta prior shape
parameters \eqn{(p, q)} for the shape parameter \eqn{\kappa}.
Only used when \code{model = "GEV"}.}

\item{years}{Numeric vector of observation years corresponding to \code{data}.
Must be the same length as \code{data} and strictly increasing.}

\item{trend}{Named list indicating which distribution parameters are
modeled as non-stationary. Must contain:
\itemize{
\item \code{location}: Logical scalar. If \code{TRUE}, the location parameter has a trend.
\item \code{scale}: Logical scalar. If \code{TRUE}, the scale parameter has a trend.
}}

\item{slice}{Numeric scalar specifying the year at which to evaluate the
quantiles or confidence intervals of a non-stationary probability distribution.
The year does not have to be an element of the \code{years} argument. Note that if
\code{trend$location} and \code{trend$scale} are both \code{FALSE}, this argument will have
no effect the output of the function.}

\item{alpha}{Numeric scalar in \eqn{[0.01, 0.1]}. The significance
level for confidence intervals or hypothesis tests. Default is 0.05.}

\item{eps}{Numeric scalar. The log-likelihood tolerance for the Regula-Falsi
convergence (default is 0.01).}
}
\value{
A list contianing the return levels and confidence intervals containing:
\itemize{
\item \code{estimates}: Estimated quantiles for each return period.
\item \code{ci_lower}: Lower bound of the confidence interval for each return period.
\item \code{ci_upper}: Upper bound of the confidence interval for each return period.
\item \code{t}: Vector of return periods \code{c(2, 5, 10, 20, 50, 100)} in years.
\item \code{slice}: The value of \code{slice} argument.
\item \code{trend}: The value of \code{trend} argument.
}
}
\description{
Calculates estimates and confidence intervals for return levels at standard
return periods (2, 5, 10, 20, 50, and 100 years) using the profile likelihood
and Regula-Falsi root‐finding method.
}
\details{
\enumerate{
\item Fits the model using \link{fit_maximum_likelihood} to obtain parameter
estimates and log‐likelihood.
\item Defines an objective function \eqn{f(y_p, p)} based on the chi-squared
distribution.
\item Iteratively brackets the root by scaling initial guesses by 0.05 until
\eqn{f(y_p, p)} changes sign.
\item Uses the Regula Falsi method to solve \eqn{f(y_p, p) = 0} for each
return-period probability.
\item Returns lower and upper confidence bounds at significance level \code{alpha}
given the return level estimates.
}
}
\note{
Although the more modern \link[stats]{optim} function is preferred over
\link[stats]{nlminb}, we use \link[stats]{nlminb} because it supports non-finite
values of the likelihood function.
}
\examples{
data <- rnorm(n = 100, mean = 100, sd = 10)
uncertainty_rfpl(data, "GLO")

}
\seealso{
\link{quantile_fast}, \link{uncertainty_bootstrap},
\link{plot_uncertainty}
}
